% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_matrix.R
\name{add_model_matrix_attribute}
\alias{add_model_matrix_attribute}
\title{Add model matrix attribute to newdata}
\usage{
add_model_matrix_attribute(mfx, newdata = NULL)
}
\arguments{
\item{mfx}{marginaleffects object}

\item{newdata}{data frame to add attributes to}
}
\description{
Add model matrix attribute to newdata
}
\keyword{internal}
