% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.comparisons}
\alias{aggregate.comparisons}
\title{Marginalize Over Unit-Level Estimates}
\usage{
\method{aggregate}{comparisons}(x, by = NULL, ...)
}
\description{
Marginalize Over Unit-Level Estimates
}
