% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_rsample2.R
\name{bootstrap_rsample2}
\alias{bootstrap_rsample2}
\title{Bootstrap rsample v2}
\usage{
bootstrap_rsample2(x, R = 1000, conf_level = 0.95, conf_type = "perc", ...)
}
\description{
Bootstrap rsample v2
}
