% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_levels.R
\name{complete_levels}
\alias{complete_levels}
\title{Create a data.frame with all factor or character levels}
\usage{
complete_levels(x, character_levels = NULL)
}
\description{
\code{model.matrix} and \code{get_predicted} break when \code{newdata} includes a factor
variable, but not all levels are present in the data. This is bad for us
because we often want to get predictions with one (or few) rows, where some
factor levels are inevitably missing.
}
\keyword{internal}
