% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{datagridcf}
\alias{datagridcf}
\title{DEPRECATED FUNCTION}
\usage{
datagridcf(..., model = NULL, newdata = NULL)
}
\description{
This function will be removed from the package in a future version. Please use this instead:
}
\details{
\preformatted{
library(marginaleffects)
mod <- lm(mpg ~ am + vs + factor(cyl) + hp, mtcars)
predictions(mod,
  newdata = datagrid(cyl = c(4, 6), grid_type = "counterfactual"))
}

Users can also create convenience shortcuts like this:
\preformatted{
datagrid_cf <- function(...) datagrid(..., grid_type = "counterfactual")
predictions(mod, newdata = datagrid_cf(cyl = c(4, 6)))
}
}
\keyword{internal}
