% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{deltamethod}
\alias{deltamethod}
\title{Estimate and Standard Error of a Non-Linear Function of Estimated Model Parameters}
\usage{
deltamethod(
  model,
  hypothesis = NULL,
  FUN = NULL,
  vcov = NULL,
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{model}{Model object or object generated by the \code{comparisons()}, \code{slopes()}, \code{predictions()}, or \code{marginalmeans()} functions.}

\item{FUN}{\code{NULL} or function.
\itemize{
\item \code{NULL} (default): hypothesis test on the model's coefficients.
\item Function which accepts a model object and returns a numeric vector or a data.frame with two columns called \code{term} and \code{estimate}. This argument can be useful when users want to conduct a hypothesis test on an arbitrary function of quantities held in a model object.
}}
}
\description{
\code{deltamethod} is a function to get a first-order approximate standard error
for a nonlinear function of a vector of random variables with known or
estimated covariance matrix. \code{\link{deltamethod}} emulates the behavior of the
excellent and well-established \link[car:deltaMethod]{car::deltaMethod} and \link[car:linearHypothesis]{car::linearHypothesis}
functions, but it supports more models, requires fewer dependencies, and
offers some convenience features like shortcuts for robust standard errors.
}
\details{
Warning: For hypothesis tests on objects produced by the \code{marginaleffects}
package, it is safer to use the \code{hypothesis} argument of the original function.
Using \code{deltamethod()} may not work in certain environments, or when called
programmatically.
}
\examples{
library(marginaleffects)
mod <- lm(mpg ~ hp + wt + factor(cyl), data = mtcars)

# When `FUN` and `hypothesis` are `NULL`, `deltamethod()` returns a data.frame of parameters
deltamethod(mod)

# Test of equality between coefficients
deltamethod(mod, hypothesis = "hp = wt")

# Non-linear function
deltamethod(mod, hypothesis = "exp(hp + wt) = 0.1")

# Robust standard errors
deltamethod(mod, hypothesis = "hp = wt", vcov = "HC3")

# b1, b2, ... shortcuts can be used to identify the position of the
# parameters of interest in the output of FUN
deltamethod(mod, hypothesis = "b2 = b3")

# term names with special characters have to be enclosed in backticks
deltamethod(mod, hypothesis = "`factor(cyl)6` = `factor(cyl)8`")

mod2 <- lm(mpg ~ hp * drat, data = mtcars)
deltamethod(mod2, hypothesis = "`hp:drat` = drat")

# predictions(), comparisons(), and slopes()
mod <- glm(am ~ hp + mpg, data = mtcars, family = binomial)
cmp <- comparisons(mod, newdata = "mean")
deltamethod(cmp, hypothesis = "b1 = b2")

mfx <- slopes(mod, newdata = "mean")
deltamethod(cmp, hypothesis = "b2 = 0.2")

pre <- predictions(mod, newdata = datagrid(hp = 110, mpg = c(30, 35)))
deltamethod(pre, hypothesis = "b1 = b2")

# The `FUN` argument can be used to compute standard errors for fitted values
mod <- glm(am ~ hp + mpg, data = mtcars, family = binomial)

f <- function(x) predict(x, type = "link", newdata = mtcars)
p <- deltamethod(mod, FUN = f)
head(p)

f <- function(x) predict(x, type = "response", newdata = mtcars)
p <- deltamethod(mod, FUN = f)
head(p)

}
