% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Fit statistical models to obtain parameter estimates}
\usage{
estimates(formula, data, model, ...)
}
\arguments{
\item{formula}{a formula specifying the model to fit}

\item{data}{a data frame containing the variables in the formula}

\item{model}{a character string specifying the model to fit. If missing, returns a list of available models and their corresponding functions.}

\item{...}{additional arguments to be passed to the model fitting function}
}
\value{
a model object. These objects can differ from model to model, but they are all supported by \code{marginaleffects} functions like \code{predictions()}, \code{slopes()}, \code{comparisons()}, and \code{hypotheses()}. These objects can also be summarized in nice tables using the \code{modelsummary} package.
}
\description{
This function offers a single point of entry for fitting many different statistical models. It provides a unified user interface for various model fitting functions, making it easier to switch between models and compare results. If the \code{model} argument is missing, the function returns a list of available models and the functions used under the hood to fit each model. If the \code{formula} argument is missing, a description of the model is printed with a list of extra arguments which can be passed to \code{estimates()} to change the model or estimation procedure.
}
\examples{
estimates(gear ~ wt, data = mtcars, model = "lm")

estimates(gear ~ wt, data = mtcars, model = "oprobit")

estimates(gear ~ wt + (1 | cyl), data = mtcars, model = "melm")

estimates(gear ~ wt, data = mtcars, model = "oprobit") |>
  avg_slopes()

}
