% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{expect_comparisons}
\alias{expect_comparisons}
\title{\code{tinytest} helper}
\usage{
expect_comparisons(object, se = TRUE, n_row = NULL, n_col = NULL, ...)
}
\description{
\code{tinytest} helper
}
\keyword{internal}
