% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{expect_marginal_means}
\alias{expect_marginal_means}
\title{\code{tinytest} helper}
\usage{
expect_marginal_means(object, se = TRUE, n_row = NULL)
}
\description{
\code{tinytest} helper
}
\keyword{internal}
