% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{expect_margins}
\alias{expect_margins}
\title{\code{tinytest} helper}
\usage{
expect_margins(
  results,
  margins_object,
  se = TRUE,
  tolerance = 1e-05,
  verbose = FALSE
)
}
\description{
\code{tinytest} helper
}
\keyword{internal}
