% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{expect_mfx}
\alias{expect_mfx}
\title{internal testing function (no validity check)}
\usage{
expect_mfx(object, prediction_type = "response", n_unique = 10)
}
\description{
internal testing function (no validity check)
}
\keyword{internal}
