% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{expect_predictions}
\alias{expect_predictions}
\title{\code{tinytest} helper}
\usage{
expect_predictions(object, se = TRUE, n_row = NULL, n_col = NULL)
}
\description{
\code{tinytest} helper
}
\keyword{internal}
