% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{extract_calling_function}
\alias{extract_calling_function}
\title{Extract calling function name from call object}
\usage{
extract_calling_function(call)
}
\arguments{
\item{call}{The call object from which to extract the function name}
}
\value{
Character string of the calling function ("comparisons", "predictions", or "hypotheses")
}
\description{
Extract calling function name from call object
}
\keyword{internal}
