% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydx.R, R/methods_betareg.R,
%   R/methods_nnet.R
\name{get_dydx}
\alias{get_dydx}
\alias{get_dydx.default}
\alias{get_dydx.betareg}
\alias{get_dydx.multinom}
\title{Get dY/dX}
\usage{
get_dydx(model, ...)

\method{get_dydx}{default}(
  model,
  fitfram,
  variable,
  variance,
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{multinom}(
  model,
  fitfram,
  variable,
  group_name,
  variance,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get dY/dX
}
