% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydx_and_se.R
\name{get_dydx_and_se}
\alias{get_dydx_and_se}
\alias{get_dydx_and_se.default}
\title{Get dY/dX}
\usage{
get_dydx_and_se(model, ...)

\method{get_dydx_and_se}{default}(
  model,
  variable,
  fitfram = insight::find_data(model),
  variance = stats::vcov(model),
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get dY/dX
}
