% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydx_categorical.R
\name{get_dydx_categorical}
\alias{get_dydx_categorical}
\title{Compute marginal effects using differences in predicted values between categorical variables (internal function)}
\usage{
get_dydx_categorical(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
Numeric vector of contrasts associated to a categorical regressor
}
\description{
Compute marginal effects using differences in predicted values between categorical variables (internal function)
}
\keyword{internal}
