% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R, R/get_dydx_stats.R,
%   R/get_dydx_nnet.R, R/get_dydx_MASS.R, R/get_dydx_betareg.R,
%   R/get_dydx_fixest.R, R/get_dydx_ivreg.R, R/get_dydx_lme4.R,
%   R/get_dydx_ordinal.R, R/get_dydx_survey.R
\name{get_gradient}
\alias{get_gradient}
\alias{get_gradient.glm}
\alias{get_gradient.lm}
\alias{get_gradient.loess}
\alias{get_gradient.multinom}
\alias{get_gradient.polr}
\alias{get_gradient.betareg}
\alias{get_gradient.fixest}
\alias{get_gradient.ivreg}
\alias{get_gradient.glmerMod}
\alias{get_gradient.lmerMod}
\alias{get_gradient.clm}
\alias{get_gradient.svyglm}
\title{Get Gradient}
\usage{
get_gradient(model, ...)

\method{get_gradient}{glm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{lm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{loess}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{multinom}(
  model,
  fitfram,
  variable,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple"
)

\method{get_gradient}{polr}(
  model,
  fitfram,
  variable,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple"
)

\method{get_gradient}{betareg}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{fixest}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{ivreg}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{glmerMod}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{lmerMod}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{clm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)

\method{get_gradient}{svyglm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple"
)
}
\description{
Get Gradient
}
