% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hypothesis.R
\name{get_hypothesis}
\alias{get_hypothesis}
\title{This function takes any input from the \code{hypothesis} argument, builds a lincom matrix and then multiplies it by the estimates}
\usage{
get_hypothesis(x, hypothesis, column, by = NULL)
}
\description{
This function takes any input from the \code{hypothesis} argument, builds a lincom matrix and then multiplies it by the estimates
}
