% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R, R/methods_betareg.R,
%   R/methods_nnet.R
\name{get_jacobian}
\alias{get_jacobian}
\alias{get_jacobian.default}
\alias{get_jacobian.betareg}
\alias{get_jacobian.multinom}
\title{Get Jacobian}
\usage{
get_jacobian(model, ...)

\method{get_jacobian}{default}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{multinom}(
  model,
  fitfram,
  variable,
  variance = NULL,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get Jacobian
}
