% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian_analytic.R, R/methods_MASS.R,
%   R/methods_brglm2.R, R/methods_geepack.R, R/methods_ivreg.R,
%   R/methods_quantreg.R, R/methods_rms.R, R/methods_stats.R,
%   R/methods_survey.R
\name{get_jacobian_analytic}
\alias{get_jacobian_analytic}
\alias{get_jacobian_analytic.default}
\alias{get_jacobian_analytic.negbin}
\alias{get_jacobian_analytic.rlm}
\alias{get_jacobian_analytic.brglmFit}
\alias{get_jacobian_analytic.geeglm}
\alias{get_jacobian_analytic.ivreg}
\alias{get_jacobian_analytic.rq}
\alias{get_jacobian_analytic.ols}
\alias{get_jacobian_analytic.lrm}
\alias{get_jacobian_analytic.lm}
\alias{get_jacobian_analytic.glm}
\alias{get_jacobian_analytic.svyglm}
\title{Get an exact analytic Jacobian when supported}
\usage{
get_jacobian_analytic(model, ...)

\method{get_jacobian_analytic}{default}(model, ...)

\method{get_jacobian_analytic}{negbin}(model, type, ...)

\method{get_jacobian_analytic}{rlm}(model, type, ...)

\method{get_jacobian_analytic}{brglmFit}(model, type, ...)

\method{get_jacobian_analytic}{geeglm}(model, type, ...)

\method{get_jacobian_analytic}{ivreg}(model, type, ...)

\method{get_jacobian_analytic}{rq}(model, type, ...)

\method{get_jacobian_analytic}{ols}(model, type, ...)

\method{get_jacobian_analytic}{lrm}(model, type, ...)

\method{get_jacobian_analytic}{lm}(model, type, ...)

\method{get_jacobian_analytic}{glm}(model, type, ...)

\method{get_jacobian_analytic}{svyglm}(model, type, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Arguments passed to model-specific methods.}
}
\value{
A numeric Jacobian matrix, or \code{NULL} when the model or estimand is
not eligible for the analytic path.
}
\description{
Get an exact analytic Jacobian when supported
}
\keyword{internal}
