% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalmeans.R
\name{get_marginalmeans}
\alias{get_marginalmeans}
\title{Workhorse function for \code{marginalmeans}}
\usage{
get_marginalmeans(model, variables, variables_grid, type, ...)
}
\arguments{
\item{model}{Model object}

\item{variables}{Categorical predictors over which to compute marginal means
(character vector). \code{NULL} calculates marginal means for all logical,
character, or factor variables in the dataset used to fit \code{model}.}

\item{variables_grid}{Categorical predictors used to construct the
prediction grid over which adjusted predictions are averaged (character
vector). \code{NULL} creates a grid with all combinations of all categorical
predictors. This grid can be very large when there are many variables and
many response levels, so it is advisable to select a limited number of
variables in the \code{variables} and \code{variables_grid} arguments.}

\item{type}{Type(s) of prediction (string or character vector). This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{...}{absorb useless arguments from other get_* workhorse functions}
}
\description{
Needs to be separate because we also need it in \code{delta_method}
}
