% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mfx_and_se.R
\name{get_mfx_and_se}
\alias{get_mfx_and_se}
\alias{get_mfx_and_se.default}
\title{Get dY/dX}
\usage{
get_mfx_and_se(model, ...)

\method{get_mfx_and_se}{default}(
  model,
  fitfram,
  variable,
  variance,
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get dY/dX
}
