% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_matrix.R, R/methods_MASS.R,
%   R/methods_brglm2.R, R/methods_fixest.R, R/methods_geepack.R,
%   R/methods_ivreg.R, R/methods_quantreg.R, R/methods_rms.R, R/methods_stats.R,
%   R/methods_survey.R
\name{get_model_matrix}
\alias{get_model_matrix}
\alias{get_model_matrix.default}
\alias{get_model_matrix.negbin}
\alias{get_model_matrix.rlm}
\alias{get_model_matrix.brglmFit}
\alias{get_model_matrix.fixest}
\alias{get_model_matrix.geeglm}
\alias{get_model_matrix.ivreg}
\alias{get_model_matrix.rq}
\alias{get_model_matrix.ols}
\alias{get_model_matrix.lrm}
\alias{get_model_matrix.lm}
\alias{get_model_matrix.glm}
\alias{get_model_matrix.svyglm}
\title{Get a named model matrix}
\usage{
get_model_matrix(model, newdata, mfx = NULL)

\method{get_model_matrix}{default}(model, newdata, mfx = NULL)

\method{get_model_matrix}{negbin}(model, newdata, mfx = NULL)

\method{get_model_matrix}{rlm}(model, newdata, mfx = NULL)

\method{get_model_matrix}{brglmFit}(model, newdata, mfx = NULL)

\method{get_model_matrix}{fixest}(model, newdata, mfx = NULL)

\method{get_model_matrix}{geeglm}(model, newdata, mfx = NULL)

\method{get_model_matrix}{ivreg}(model, newdata, mfx = NULL)

\method{get_model_matrix}{rq}(model, newdata, mfx = NULL)

\method{get_model_matrix}{ols}(model, newdata, mfx = NULL)

\method{get_model_matrix}{lrm}(model, newdata, mfx = NULL)

\method{get_model_matrix}{lm}(model, newdata, mfx = NULL)

\method{get_model_matrix}{glm}(model, newdata, mfx = NULL)

\method{get_model_matrix}{svyglm}(model, newdata, mfx = NULL)
}
\arguments{
\item{model}{Model object}

\item{newdata}{Grid of predictor values at which we evaluate the slopes.
\itemize{
\item Warning: When \code{newdata} is \code{NULL}, the data is retrieved using \code{\link[=get_modeldata]{get_modeldata()}}, which may have to extract it from the modeling environment. This can produce unexpected results if the data has been modified, or when called inside \code{lapply()}, Shiny apps, or nested functions. To avoid potential issues, consider using \code{\link[=set_modeldata]{set_modeldata()}} to attach the training data to the model object explicitly.
\item \code{NULL} (default): Unit-level slopes for each observed value in the dataset (empirical distribution). The dataset is retrieved using \code{\link[=get_modeldata]{get_modeldata()}}.
\item \code{\link[=datagrid]{datagrid()}} call to specify a custom grid of regressors. For example:
\itemize{
\item \code{newdata = datagrid(cyl = c(4, 6))}: \code{cyl} variable equal to 4 and 6 and other regressors fixed at their means or modes.
\item See the Examples section and the \code{\link[=datagrid]{datagrid()}} documentation.
}
\item \code{\link[=subset]{subset()}} call with a single argument to select a subset of the dataset used to fit the model, ex: \code{newdata = subset(treatment == 1)}
\item \code{\link[dplyr:filter]{dplyr::filter()}} call with a single argument to select a subset of the dataset used to fit the model, ex: \code{newdata = filter(treatment == 1)}
\item string:
\itemize{
\item "mean": Slopes evaluated when each predictor is held at its mean or mode.
\item "median": Slopes evaluated when each predictor is held at its median or mode.
\item "balanced": Slopes evaluated on a balanced grid with every combination of categories and numeric variables held at their means.
\item "tukey": Slopes evaluated at Tukey's 5 numbers.
\item "grid": Slopes evaluated on a grid of representative numbers (Tukey's 5 numbers and unique values of categorical predictors).
}
}}
}
\description{
Get a named model matrix
}
\keyword{internal}
