% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modeldata.R
\name{get_modeldata}
\alias{get_modeldata}
\title{Get the dataset used to fit a model}
\usage{
get_modeldata(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are passed to the \code{predict()} method
supplied by the modeling package.These arguments are particularly useful
for mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?slopes} documentation for a non-exhaustive list of available
arguments.}
}
\value{
A data frame of the original data used to fit the model, or \code{NULL} if not available.
}
\description{
Mostly for internal use, but can be useful because the output is consistent across model classes.
}
