% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_delta.R, R/methods_nnet.R
\name{get_se_delta}
\alias{get_se_delta}
\alias{get_se_delta.default}
\alias{get_se_delta.multinom}
\title{Get Jacobian}
\usage{
get_se_delta(model, ...)

\method{get_se_delta}{default}(
  model,
  fitfram,
  variable,
  variance,
  group_name,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_se_delta}{multinom}(
  model,
  fitfram,
  variable,
  variance = NULL,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get Jacobian
}
