% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R, R/methods_betareg.R,
%   R/methods_nnet.R
\name{get_semfx_delta}
\alias{get_semfx_delta}
\alias{get_semfx_delta.default}
\alias{get_semfx_delta.betareg}
\alias{get_semfx_delta.multinom}
\title{Get Jacobian}
\usage{
get_semfx_delta(model, ...)

\method{get_semfx_delta}{default}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_semfx_delta}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_semfx_delta}{multinom}(
  model,
  fitfram,
  variable,
  variance = NULL,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get Jacobian
}
