% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcov.R
\name{get_varcov_args}
\alias{get_varcov_args}
\title{Take a \code{modelsummary()} style \code{vcov} argument and convert it to
\code{insight::get_varcov()}}
\usage{
get_varcov_args(model, vcov)
}
\description{
Take a \code{modelsummary()} style \code{vcov} argument and convert it to
\code{insight::get_varcov()}
}
\keyword{internal}
