% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.marginaleffects}
\alias{glance.marginaleffects}
\title{Glance at key characteristics of an object}
\usage{
\method{glance}{marginaleffects}(x, ...)
}
\description{
Glance at key characteristics of an object
}
