% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hush.R
\name{hush}
\alias{hush}
\title{Execute code silently}
\usage{
hush(code)
}
\description{
Execute code silently
}
\keyword{Do}
\keyword{avoid}
\keyword{conflict}
\keyword{export}
\keyword{internal}
\keyword{modelsummary}
\keyword{not}
\keyword{to}
\keyword{with}
