% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference and uncertainy estimates for \code{marginaleffects} objects \link{EXPERIMENTAL}}
\usage{
inference(model, method = "simulation", iter = 1000, ...)
}
\arguments{
\item{model}{Model object}

\item{method}{String: "simulation" or "delta". See below for references and details.}

\item{iter}{Number of simulations.}

\item{...}{Other arguments are ignored.}
}
\value{
A "decorated" model with additional information on how to conduct inference and compute uncertainty estimates. This decorated model can be used in any of the \code{marginaleffects} functions, as all other supported models.
}
\description{
This function is experimental. The user interface may change, or the functionality may migrate to arguments of individual \code{marginaleffects} functions.
}
\details{
When \code{method="simulation"}, we conduct simulation-based inference following the method discussed in Krinsky & Robb (1986):
\enumerate{
\item Draw \code{iter} sets of simulated coefficients from a multivariate normal distribution with mean equal to the original model's estimated coefficients and variance equal to the model's variance-covariance matrix (classical, "HC3", or other).
\item Use the \code{iter} sets of coefficients to compute \code{iter} sets of estimands: predictions, comparisons, or slopes.
\item Take quantiles of the resulting distribution of estimands to obtain a confidence interval and the standard deviation of simulated estimates to estimate the standard error.
}
}
\section{References}{

\itemize{
\item Krinsky, I., and A. L. Robb. 1986. “On Approximating the Statistical Properties of Elasticities.” Review of Economics and Statistics 68 (4): 715–9.
\item Dowd, Bryan E., William H. Greene, and Edward C. Norton. "Computation of standard errors." Health services research 49.2 (2014): 731-750.
\item King, Gary, Michael Tomz, and Jason Wittenberg. "Making the most of statistical analyses: Improving interpretation and presentation." American journal of political science (2000): 347-361
}
}

\examples{
library(marginaleffects)
library(magrittr)
mod <- glm(vs ~ hp * wt + factor(gear), data = mtcars, family = binomial)

mod \%>\% inference() \%>\%
        avg_predictions(by = "gear")

mod \%>\% inference() \%>\%
        slopes() \%>\%
        head()

mod \%>\% inference() \%>\%
        avg_slopes() \%>\%
        posterior_draws("rvar")

}
