% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences.R
\name{inferences}
\alias{inferences}
\title{EXPERIMENTAL Uncertainty Estimates for \code{marginaleffects} Objects}
\usage{
inferences(model, method = "simulation", R = 1000, ...)
}
\arguments{
\item{model}{Model object}

\item{method}{String: "simulation", "boot", or "delta". See below for references and details.}

\item{R}{Number of simulations.}

\item{...}{Other arguments are ignored.}
}
\value{
A "decorated" model with additional information on how to conduct inference and compute uncertainty estimates. This decorated model can then fed to any of the \code{marginaleffects} functions, as one would for any other supported models.
}
\description{
Warning: This function is experimental. It may be renamed, the user interface may change, or the functionality may migrate to arguments in other \code{marginaleffects} functions.

Apply this function to a model object to change the inferential strategy used to compute uncertainty estimates: delta method or simulation-based inference.
}
\details{
When \code{method="simulation"}, we conduct simulation-based inference following the method discussed in Krinsky & Robb (1986):
\enumerate{
\item Draw \code{R} sets of simulated coefficients from a multivariate normal distribution with mean equal to the original model's estimated coefficients and variance equal to the model's variance-covariance matrix (classical, "HC3", or other).
\item Use the \code{R} sets of coefficients to compute \code{R} sets of estimands: predictions, comparisons, or slopes.
\item Take quantiles of the resulting distribution of estimands to obtain a confidence interval and the standard deviation of simulated estimates to estimate the standard error.
}
}
\section{References}{


Krinsky, I., and A. L. Robb. 1986. “On Approximating the Statistical Properties of Elasticities.” Review of Economics and Statistics 68 (4): 715–9.

King, Gary, Michael Tomz, and Jason Wittenberg. "Making the most of statistical analyses: Improving interpretation and presentation." American journal of political science (2000): 347-361

Dowd, Bryan E., William H. Greene, and Edward C. Norton. "Computation of standard errors." Health services research 49.2 (2014): 731-750.
}

\examples{
library(marginaleffects)
library(magrittr)
mod <- glm(vs ~ hp * wt + factor(gear), data = mtcars, family = binomial)

mod \%>\%
    inferences() \%>\%
    avg_predictions(by = "gear")

mod \%>\%
    inferences() \%>\%
    slopes() \%>\%
    head()

mod \%>\%
    inferences() \%>\%
    avg_slopes() \%>\%
    posterior_draws("rvar")

}
