% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_rsample2.R
\name{inferences_rsample}
\alias{inferences_rsample}
\title{Bootstrap rsample v2}
\usage{
inferences_rsample(x, R = 1000, conf_level = 0.95, conf_type = "perc", ...)
}
\description{
Bootstrap rsample v2
}
