% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences_simulation.R
\name{inferences_simulation}
\alias{inferences_simulation}
\title{Simulation-based inference}
\usage{
inferences_simulation(x, R = 1000, conf_level = 0.95, ...)
}
\description{
Simulation-based inference
}
