% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.marginaleffects}
\alias{knit_print.marginaleffects}
\title{Print a marginaleffects object in knitr}
\usage{
knit_print.marginaleffects(x, ...)
}
\value{
A string with class \code{knit_asis} to be printed in Rmarkdown or Quarto documents.
}
\description{
Print a marginaleffects object in knitr
}
\keyword{internal}
