% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{marginal_means}
\alias{marginal_means}
\title{Deprecated function}
\usage{
marginal_means(
  model,
  variables = NULL,
  newdata = NULL,
  vcov = TRUE,
  conf_level = 0.95,
  type = NULL,
  transform = NULL,
  cross = FALSE,
  hypothesis = NULL,
  equivalence = NULL,
  p_adjust = NULL,
  df = Inf,
  wts = "equal",
  by = NULL,
  numderiv = "fdforward",
  ...
)
}
\description{
Deprecated function
}
\keyword{internal}
