% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_or_mode.R
\name{mean_or_mode}
\alias{mean_or_mode}
\alias{mean_or_mode.default}
\alias{mean_or_mode.character}
\alias{mean_or_mode.factor}
\alias{mean_or_mode.logical}
\alias{mean_or_mode.data.frame}
\title{Compute the mode or mean of \code{x}}
\usage{
mean_or_mode(x)

\method{mean_or_mode}{default}(x)

\method{mean_or_mode}{character}(x)

\method{mean_or_mode}{factor}(x)

\method{mean_or_mode}{logical}(x)

\method{mean_or_mode}{data.frame}(x)
}
\arguments{
\item{x}{extract the mean or the mode of vector or data.frame x depending on its type}
}
\value{
numeric vector
}
\description{
Compute the mode or mean of \code{x}
}
\keyword{internal}
