% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R
\name{median_or_mode}
\alias{median_or_mode}
\alias{median_or_mode.default}
\alias{median_or_mode.factor}
\alias{median_or_mode.logical}
\alias{median_or_mode.data.frame}
\title{Get typical values from data.frames or vectors}
\usage{
median_or_mode(x)

\method{median_or_mode}{default}(x)

\method{median_or_mode}{factor}(x)

\method{median_or_mode}{logical}(x)

\method{median_or_mode}{data.frame}(x)
}
\description{
Get typical values from data.frames or vectors
}
