% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{meffects}
\alias{meffects}
\title{\code{meffects()} is an alias to \code{slopes()}}
\usage{
meffects(
  model,
  newdata = NULL,
  variables = NULL,
  vcov = TRUE,
  conf_level = 0.95,
  type = NULL,
  slope = "dydx",
  by = NULL,
  wts = NULL,
  hypothesis = NULL,
  eps = NULL,
  ...
)
}
\description{
This alias is kept for backward compatibility and because some users may prefer that name.
}
\keyword{internal}
