% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfx.R
\name{mfx}
\alias{mfx}
\title{compute marginal effects estimates using numerical derivatives}
\usage{
mfx(
  model,
  variables = NULL,
  fitfram = NULL,
  group_names = NULL,
  variance = vcov(model)
)
}
\description{
compute marginal effects estimates using numerical derivatives
}
