% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual.R
\name{Mode}
\alias{Mode}
\title{find mode, preserve type, and pick an arbitrary value when multi-modal
https://stackoverflow.com/a/8189441/342331}
\usage{
Mode(x)
}
\description{
find mode, preserve type, and pick an arbitrary value when multi-modal
https://stackoverflow.com/a/8189441/342331
}
