% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Pad newdata with factor levels and merge with original data}
\usage{
pad(model, newdata)
}
\arguments{
\item{model}{Model object to check for mlogit class}

\item{newdata}{Data frame to pad}
}
\description{
\code{model.matrix} breaks when \code{newdata} includes a factor
variable, but not all levels are present in the data. This is bad for us
because we often want to get predictions with one (or few) rows, where some
factor levels are inevitably missing.
}
\keyword{internal}
