% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slopes}
\alias{plot.slopes}
\title{Point-range plot of average marginal effects}
\usage{
\method{plot}{slopes}(x, conf_level = 0.95, ...)
}
\value{
A \code{ggplot2} object
}
\description{
Uses the \code{ggplot2} package to draw a point-range plot of the average marginal effects computed by \code{tidy}.
}
\examples{
mod <- glm(am ~ hp + wt, data = mtcars)
mfx <- slopes(mod)
plot(mfx)

}
\seealso{
Other plot: 
\code{\link{plot_cap}()},
\code{\link{plot_cco}()},
\code{\link{plot_cme}()}
}
\concept{plot}
