% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_agg}
\alias{plot_agg}
\title{Plot aggregated (average) comparisons and slopes}
\usage{
plot_agg(x, ...)
}
\arguments{
\item{x}{object produced by the \code{comparisons()} or \code{slopes()} object.}

\item{...}{additional arguments are passed to the \code{aggregate} function (e.g., \code{conf_level}).}
}
\value{
A \code{ggplot2} object
}
\description{
\code{x} is passed to \code{aggregate()} and the result is drawn as a point-range plot using \code{ggplot2}.
}
\examples{
mod <- glm(am ~ hp + wt, data = mtcars)
mfx <- slopes(mod)
plot_agg(mfx)


mod <- glm(am ~ hp + factor(gear), data = mtcars)
cmp <- comparisons(mod)
plot_agg(cmp)
}
\seealso{
Other plot: 
\code{\link{plot_cap}()},
\code{\link{plot_cco}()},
\code{\link{plot_cme}()}
}
\concept{plot}
