% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cap.R
\name{plot_cap}
\alias{plot_cap}
\title{Plot Conditional Adjusted Predictions}
\usage{
plot_cap(
  model,
  condition = NULL,
  type = "response",
  vcov = NULL,
  conf_level = 0.95,
  transform_post = NULL,
  draw = TRUE,
  ...
)
}
\arguments{
\item{model}{Model object}

\item{condition}{character vector or named list of length smaller than 4. Character vectors must be the names of the predictor variables to display. The names of the list must The first element is displayed on the x-axis. The second element determines the colors. The third element creates facets. Other variables are held at their means or modes. Lists can include these types of values:
\itemize{
\item Numeric vector
\item Function which returns a numeric vector or a set of unique categorical values
\item Shortcut strings for common reference values: "minmax", "quartile", "threenum"
}}

\item{transform_post}{(experimental) A function applied to unit-level adjusted predictions and confidence intervals just before the function returns results. For bayesian models, this function is applied to individual draws from the posterior distribution, before computing summaries.}

\item{draw}{\code{TRUE} returns a \code{ggplot2} plot. \code{FALSE} returns a \code{data.frame} of the underlying data.}
}
\value{
A \code{ggplot2} object
}
\description{
This function plots adjusted predictions (y-axis) against values of one or
more predictors (x-axis and colors).
}
\examples{
mod <- lm(mpg ~ hp + wt, data = mtcars)
plot_cap(mod, condition = "wt")

mod <- lm(mpg ~ hp * wt * am, data = mtcars)
plot_cap(mod, condition = c("hp", "wt"))

plot_cap(mod, condition = list("hp", wt = "threenum"))

plot_cap(mod, condition = list("hp", wt = range))

}
\seealso{
Other plot: 
\code{\link{plot_agg}()},
\code{\link{plot_cco}()},
\code{\link{plot_cme}()}
}
\concept{plot}
