% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cco.R
\name{plot_cco}
\alias{plot_cco}
\title{Plot Conditional Contrasts}
\usage{
plot_cco(
  model,
  effect = NULL,
  condition = NULL,
  type = "response",
  vcov = NULL,
  conf_level = 0.95,
  transform_pre = "difference",
  transform_post = NULL,
  draw = TRUE,
  ...
)
}
\arguments{
\item{model}{Model object}

\item{effect}{Name of the variable whose contrast we want to plot on the y-axis}

\item{condition}{character vector or named list of length smaller than 3. Character vectors must be the names of the predictor variables to display. The names of the list must The first element is displayed on the x-axis. The second element determines the colors. The third element creates facets. Unspecified variables are held at their means or modes. Lists can include these types of values (see Examples section below):
\itemize{
\item Numeric vector
\item Function which returns a numeric vector or a set of unique categorical values
\item Shortcut strings for common reference values: "minmax", "quartile", "threenum"
}}

\item{conf_level}{numeric value between 0 and 1. Confidence level to use to build a confidence interval. The default \code{NULL} uses the \code{conf_level} value used in the original call to \code{slopes()}.}

\item{transform_pre}{string or function. How should pairs of adjusted predictions be contrasted?
\itemize{
\item string: shortcuts to common contrast functions.
\itemize{
\item Supported shortcuts strings: difference, differenceavg, differenceavgwts, dydx, eyex, eydx, dyex, dydxavg, eyexavg, eydxavg, dyexavg, dydxavgwts, eyexavgwts, eydxavgwts, dyexavgwts, ratio, ratioavg, ratioavgwts, lnratio, lnratioavg, lnratioavgwts, lnor, lnoravg, lnoravgwts, expdydx, expdydxavg, expdydxavgwts
\item See the Transformations section below for definitions of each transformation.
}
\item function: accept two equal-length numeric vectors of adjusted predictions (\code{hi} and \code{lo}) and returns a vector of contrasts of the same length, or a unique numeric value.
\itemize{
\item See the Transformations section below for examples of valid functions.
}
}}

\item{transform_post}{string or function. Transformation applied to unit-level estimates and confidence intervals just before the function returns results. Functions must accept a vector and return a vector of the same length. Support string shortcuts: "exp", "ln"}

\item{draw}{\code{TRUE} returns a \code{ggplot2} plot. \code{FALSE} returns a \code{data.frame} of the underlying data.}
}
\value{
A \code{ggplot2} object
}
\description{
This function plots contrasts (y-axis) against values of predictor(s)
variable(s) (x-axis and colors). This is especially useful in models with
interactions, where the values of contrasts depend on the values of
"condition" variables.
}
\examples{
mod <- lm(mpg ~ hp * drat * factor(am), data = mtcars)

plot_cco(mod, effect = "hp", condition = "drat")

plot_cco(mod, effect = "hp", condition = c("drat", "am"))

plot_cco(mod, effect = "hp", condition = list("am", "drat" = 3:5))

plot_cco(mod, effect = "am", condition = list("hp", "drat" = range))

}
\seealso{
Other plot: 
\code{\link{plot.slopes}()},
\code{\link{plot_cap}()},
\code{\link{plot_cme}()}
}
\concept{plot}
