% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cme.R
\name{plot_cme}
\alias{plot_cme}
\title{Conditional marginal effects plot}
\usage{
plot_cme(
  model,
  effect,
  condition,
  conf.int = TRUE,
  conf.level = 0.95,
  draw = TRUE
)
}
\description{
Conditional marginal effects plot
}
