% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmm.R
\name{plot_cmm}
\alias{plot_cmm}
\title{Conditional marginal means plot}
\usage{
plot_cmm(
  model,
  condition,
  predict_type = "response",
  conf.int = TRUE,
  conf.level = 0.95,
  draw = TRUE
)
}
\arguments{
\item{model}{Model object}

\item{condition}{String or vector of two strings. The first is a variable
name to be displayed on the x-axis. The second is a variable whose values
will be displayed in different colors.}

\item{predict_type}{Type(s) of prediction as string or vector This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{conf.int}{Logical indicating whether or not to include a confidence interval.}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int=TRUE}. Must be strictly greater than 0 and less than 1. Defaults
to 0.95, which corresponds to a 95 percent confidence interval.}

\item{draw}{\code{TRUE} returns a \code{ggplot2} plot. \code{FALSE} returns a \code{data.frame} of the underlying data.}
}
\value{
A \code{ggplot2} object
}
\description{
This function plots the marginal mean of the outcome variable (y-axis)
against values of one or more predictors.
}
