% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draws.R
\name{posterior_draws}
\alias{posterior_draws}
\title{\code{posteriordraws()} is an alias to \code{get_draws()}}
\usage{
posterior_draws(x, shape = "long")
}
\description{
This alias is kept for backward compatibility and because some users may prefer that name.
}
\keyword{internal}
