% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.marginaleffects}
\alias{print.marginaleffects}
\title{Print \code{marginaleffects} objects}
\usage{
\method{print}{marginaleffects}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  topn = getOption("marginaleffects_print_topn", default = 5),
  nrows = getOption("marginaleffects_print_nrows", default = 30),
  ncols = getOption("marginaleffects_print_ncols", default = 30),
  style = getOption("marginaleffects_print_style", default = "summary"),
  ...
)
}
\arguments{
\item{x}{An object produced by one of the \code{\link{marginaleffects}} package functions.}

\item{digits}{The number of digits to display.}

\item{topn}{The number of rows to be printed from the beginning and end of tables with more than \code{nrows} rows.}

\item{nrows}{The number of rows which will be printed before truncation.}

\item{ncols}{The maximum number of column names to display at the bottom of the printed output.}

\item{style}{"summary" or "data.frame"}

\item{...}{Other arguments are currently ignored.}
}
\description{
This function controls the text printed to the \code{R} console when executing one of the core \code{marginaleffects} functions.
}
