% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.marginalmeans.summary}
\alias{print.marginalmeans.summary}
\title{Print a \code{marginalmeans} summary}
\usage{
\method{print}{marginalmeans.summary}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object produced by the \code{marginaleffects} function.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
Printed summary of a \code{marginalmeans} object
}
\description{
Print a \code{marginalmeans} summary
}
