% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{prune.marginaleffects}
\alias{prune.marginaleffects}
\title{Prune marginaleffects objects to reduce memory usage}
\usage{
\method{prune}{marginaleffects}(tree, ...)
}
\arguments{
\item{tree}{A marginaleffects object (predictions, comparisons, slopes, or hypotheses)}

\item{...}{Unused}
}
\value{
A pruned marginaleffects object
}
\description{
Remove large attributes from marginaleffects objects to reduce memory usage.
Warning: This will disable many useful post-processing features of \code{marginaleffects}
}
\details{
...
}
