% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune.comparisons}
\alias{prune.comparisons}
\alias{prune.predictions}
\alias{prune.hypotheses}
\alias{prune.slopes}
\title{Prune marginaleffects objects to reduce memory usage}
\usage{
\method{prune}{comparisons}(tree, ...)

\method{prune}{predictions}(tree, ...)

\method{prune}{hypotheses}(tree, ...)

\method{prune}{slopes}(tree, ...)
}
\arguments{
\item{tree}{A marginaleffects object (predictions, comparisons, slopes, or hypotheses)}

\item{...}{Unused}
}
\value{
A pruned marginaleffects object
}
\description{
Remove large attributes from marginaleffects objects to reduce memory usage.
Warning: This will disable many useful post-processing features of \code{marginaleffects}
}
\details{
...
}
