% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glance}
\alias{tidy}
\title{Objects exported from other packages}
\seealso{
Other summary: 
\code{\link{glance.marginaleffects}()},
\code{\link{summary.comparisons}()},
\code{\link{summary.marginaleffects}()},
\code{\link{summary.marginalmeans}()},
\code{\link{summary.predictions}()},
\code{\link{tidy.comparisons}()},
\code{\link{tidy.deltamethod}()},
\code{\link{tidy.marginaleffects}()},
\code{\link{tidy.marginalmeans}()},
\code{\link{tidy.predictions}()}
}
\concept{summary}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
}}

