% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_model.R, R/methods_DirichletReg.R,
%   R/methods_aod.R, R/methods_betareg.R, R/methods_biglm.R, R/methods_brms.R,
%   R/methods_dbarts.R, R/methods_fixest.R, R/methods_gamlss.R,
%   R/methods_glmmTMB.R, R/methods_lme4.R, R/methods_mclogit.R,
%   R/methods_mlogit.R, R/methods_mlr3.R, R/methods_mmrm.R, R/methods_ordinal.R,
%   R/methods_plm.R, R/methods_quantreg.R, R/methods_rms.R, R/methods_survey.R,
%   R/methods_survival.R, R/methods_svyVGAM.R
\name{sanitize_model_specific}
\alias{sanitize_model_specific}
\alias{sanitize_model_specific.default}
\alias{sanitize_model_specific.DirichletRegModel}
\alias{sanitize_model_specific.glimML}
\alias{sanitize_model_specific.betareg}
\alias{sanitize_model_specific.biglm}
\alias{sanitize_model_specific.brmsfit}
\alias{sanitize_model_specific.bart}
\alias{sanitize_model_specific.fixest}
\alias{sanitize_model_specific.gamlss}
\alias{sanitize_model_specific.glmmTMB}
\alias{sanitize_model_specific.merMod}
\alias{sanitize_model_specific.mblogit}
\alias{sanitize_model_specific.mlogit}
\alias{sanitize_model_specific.Learner}
\alias{sanitize_model_specific.mmrm}
\alias{sanitize_model_specific.clm}
\alias{sanitize_model_specific.clmm2}
\alias{sanitize_model_specific.plm}
\alias{sanitize_model_specific.rqs}
\alias{sanitize_model_specific.rms}
\alias{sanitize_model_specific.orm}
\alias{sanitize_model_specific.lrm}
\alias{sanitize_model_specific.ols}
\alias{sanitize_model_specific.svyolr}
\alias{sanitize_model_specific.svyglm}
\alias{sanitize_model_specific.coxph}
\alias{sanitize_model_specific.svy_vglm}
\title{Method to raise model-specific warnings and errors}
\usage{
sanitize_model_specific(model, ...)

\method{sanitize_model_specific}{default}(
  model,
  vcov = NULL,
  calling_function = "marginaleffects",
  ...
)

\method{sanitize_model_specific}{DirichletRegModel}(model, calling_function = "marginaleffects", ...)

\method{sanitize_model_specific}{glimML}(model, ...)

\method{sanitize_model_specific}{betareg}(model, ...)

\method{sanitize_model_specific}{biglm}(model, vcov = NULL, ...)

\method{sanitize_model_specific}{brmsfit}(model, ...)

\method{sanitize_model_specific}{bart}(model, calling_function, ...)

\method{sanitize_model_specific}{fixest}(
  model,
  vcov = TRUE,
  calling_function = "predictions",
  ...
)

\method{sanitize_model_specific}{gamlss}(model, calling_function, ...)

\method{sanitize_model_specific}{glmmTMB}(model, vcov = TRUE, re.form, ...)

\method{sanitize_model_specific}{merMod}(model, re.form, vcov = TRUE, ...)

\method{sanitize_model_specific}{mblogit}(model, calling_function = "marginaleffects", ...)

\method{sanitize_model_specific}{mlogit}(model, calling_function = NULL, ...)

\method{sanitize_model_specific}{Learner}(model, calling_function, ...)

\method{sanitize_model_specific}{mmrm}(model, ...)

\method{sanitize_model_specific}{clm}(model, ...)

\method{sanitize_model_specific}{clmm2}(model, ...)

\method{sanitize_model_specific}{plm}(model, ...)

\method{sanitize_model_specific}{rqs}(model, ...)

\method{sanitize_model_specific}{rms}(model, ...)

\method{sanitize_model_specific}{orm}(model, ...)

\method{sanitize_model_specific}{lrm}(model, ...)

\method{sanitize_model_specific}{ols}(model, ...)

\method{sanitize_model_specific}{svyolr}(model, calling_function = NULL, ...)

\method{sanitize_model_specific}{svyglm}(model, calling_function = NULL, ...)

\method{sanitize_model_specific}{coxph}(model, vcov, ...)

\method{sanitize_model_specific}{svy_vglm}(model, calling_function = NULL, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are passed to the \code{predict()} method
supplied by the modeling package.These arguments are particularly useful
for mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?slopes} documentation for a non-exhaustive list of available
arguments.}

\item{vcov}{Type of uncertainty estimates to report (e.g., for robust standard errors). Acceptable values:
\itemize{
\item FALSE: Do not compute standard errors. This can speed up computation considerably.
\item TRUE: Unit-level standard errors using the default \code{vcov(model)} variance-covariance matrix.
\item String which indicates the kind of uncertainty estimates to return.
\itemize{
\item Heteroskedasticity-consistent: \code{"HC"}, \code{"HC0"}, \code{"HC1"}, \code{"HC2"}, \code{"HC3"}, \code{"HC4"}, \code{"HC4m"}, \code{"HC5"}. See \code{?sandwich::vcovHC}
\item Heteroskedasticity and autocorrelation consistent: \code{"HAC"}
\item Unconditional: \code{"unconditional"} accounts for sampling variation in
the empirical covariate distribution for averaged or aggregated
predictions, comparisons, and slopes. Hypotheses applied directly to
unit-level effects are rejected. Use
\code{vcovUnconditional(cluster = ~cluster)} for
one-way clustered unconditional inference.
\item Mixed-Models degrees of freedom: "satterthwaite", "kenward-roger"
\item Other: \code{"NeweyWest"}, \code{"KernHAC"}, \code{"OPG"}. See the \code{sandwich} package documentation.
\item \code{"rsample"}, \code{"boot"}, \code{"fwb"}, or \code{"simulation"}: forward the result to \code{inferences()} using that method.
}
\item One-sided formula which indicates the name of cluster variables (e.g., \code{~unit_id}). This formula is passed to the \code{cluster} argument of the \code{sandwich::vcovCL} function.
\item Square covariance matrix
\item Function which returns a covariance matrix (e.g., \code{stats::vcov(model)})
}}
}
\value{
A warning, an error, or nothing
}
\description{
Method to raise model-specific warnings and errors
}
\keyword{internal}
