% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_type.R
\name{sanity_type}
\alias{sanity_type}
\title{check type sanity}
\usage{
sanity_type(model, type)
}
\arguments{
\item{model}{model object}

\item{type}{character vector}
}
\value{
Named vector where value is the Base R \code{type} and name is the
insight::get_predicted \code{predict}
}
\description{
check type sanity
}
