% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_coef.R, R/methods_AER.R,
%   R/methods_DirichletReg.R, R/methods_MASS.R, R/methods_Rchoice.R,
%   R/methods_aaa.R, R/methods_afex.R, R/methods_aod.R, R/methods_betareg.R,
%   R/methods_nnet.R, R/methods_causalgenerics.R, R/methods_crch.R,
%   R/methods_dataframe.R, R/methods_flexsurv.R, R/methods_gamlss.R,
%   R/methods_glmmTMB.R, R/methods_glmtoolbox.R, R/methods_glmx.R,
%   R/methods_lme4.R, R/methods_mlm.R, R/methods_nestedLogit.R,
%   R/methods_nlme.R, R/methods_pscl.R, R/methods_robustlmm.R,
%   R/methods_rstpm2.R, R/methods_sampleSelection.R, R/methods_scam.R,
%   R/methods_survey.R, R/methods_survival.R, R/methods_svyVGAM.R,
%   R/methods_systemfit.R, R/methods_tidymodels.R
\name{set_coef}
\alias{set_coef}
\alias{set_coef.default}
\alias{set_coef.tobit}
\alias{set_coef.DirichletRegModel}
\alias{set_coef.polr}
\alias{set_coef.glmmPQL}
\alias{set_coef.hetprob}
\alias{set_coef.ivpml}
\alias{set_coef.glm}
\alias{set_coef.lm}
\alias{set_coef.nls}
\alias{set_coef.afex_aov}
\alias{set_coef.glimML}
\alias{set_coef.betareg}
\alias{set_coef.multinom}
\alias{set_coef.ipw}
\alias{set_coef.crch}
\alias{set_coef.hxlr}
\alias{set_coef.data.frame}
\alias{set_coef.flexsurvreg}
\alias{set_coef.gamlss}
\alias{set_coef.glmmTMB}
\alias{set_coef.glmgee}
\alias{set_coef.gnm}
\alias{set_coef.glmx}
\alias{set_coef.merMod}
\alias{set_coef.lmerModLmerTest}
\alias{set_coef.lmerMod}
\alias{set_coef.mlm}
\alias{set_coef.nestedLogit}
\alias{set_coef.lme}
\alias{set_coef.hurdle}
\alias{set_coef.zeroinfl}
\alias{set_coef.rlmerMod}
\alias{set_coef.stpm2}
\alias{set_coef.pstpm2}
\alias{set_coef.gsm}
\alias{set_coef.aft}
\alias{set_coef.selection}
\alias{set_coef.scam}
\alias{set_coef.svyolr}
\alias{set_coef.survreg}
\alias{set_coef.svy_vglm}
\alias{set_coef.systemfit}
\alias{set_coef.model_fit}
\alias{set_coef.workflow}
\title{Internal function to set coefficients}
\usage{
set_coef(model, coefs, ...)

\method{set_coef}{default}(model, coefs, ...)

\method{set_coef}{tobit}(model, coefs, ...)

\method{set_coef}{DirichletRegModel}(model, coefs, ...)

\method{set_coef}{polr}(model, coefs, ...)

\method{set_coef}{glmmPQL}(model, coefs, ...)

\method{set_coef}{hetprob}(model, coefs, ...)

\method{set_coef}{ivpml}(model, coefs, ...)

\method{set_coef}{glm}(model, coefs, ...)

\method{set_coef}{lm}(model, coefs, ...)

\method{set_coef}{nls}(model, coefs, ...)

\method{set_coef}{afex_aov}(model, coefs, ...)

\method{set_coef}{glimML}(model, coefs, ...)

\method{set_coef}{betareg}(model, coefs, ...)

\method{set_coef}{multinom}(model, coefs, ...)

\method{set_coef}{ipw}(model, coefs, ...)

\method{set_coef}{crch}(model, coefs, ...)

\method{set_coef}{hxlr}(model, coefs, ...)

\method{set_coef}{data.frame}(model, coefs, ...)

\method{set_coef}{flexsurvreg}(model, coefs, ...)

\method{set_coef}{gamlss}(model, coefs, ...)

\method{set_coef}{glmmTMB}(model, coefs, ...)

\method{set_coef}{glmgee}(model, coefs, ...)

\method{set_coef}{gnm}(model, coefs, ...)

\method{set_coef}{glmx}(model, coefs, ...)

\method{set_coef}{merMod}(model, coefs, ...)

\method{set_coef}{lmerModLmerTest}(model, coefs, ...)

\method{set_coef}{lmerMod}(model, coefs, ...)

\method{set_coef}{mlm}(model, coefs, ...)

\method{set_coef}{nestedLogit}(model, coefs, ...)

\method{set_coef}{lme}(model, coefs, ...)

\method{set_coef}{hurdle}(model, coefs, ...)

\method{set_coef}{zeroinfl}(model, coefs, ...)

\method{set_coef}{rlmerMod}(model, coefs, ...)

\method{set_coef}{stpm2}(model, coefs, ...)

\method{set_coef}{pstpm2}(model, coefs, ...)

\method{set_coef}{gsm}(model, coefs, ...)

\method{set_coef}{aft}(model, coefs, ...)

\method{set_coef}{selection}(model, coefs, ...)

\method{set_coef}{scam}(model, coefs, ...)

\method{set_coef}{svyolr}(model, coefs, ...)

\method{set_coef}{survreg}(model, coefs, ...)

\method{set_coef}{svy_vglm}(model, coefs, ...)

\method{set_coef}{systemfit}(model, coefs, ...)

\method{set_coef}{model_fit}(model, coefs, ...)

\method{set_coef}{workflow}(model, coefs, ...)
}
\arguments{
\item{model}{object to modify}

\item{coefs}{vector of coefficients to insert in the model object}
}
\value{
Model object of the same class as the \code{model} argument, but with
different stored coefficients.
}
\description{
Set the coefficients in a model to different values and return the modified object (internal function)
}
\details{
To compute the variance of marginal effects we need to take the
Jacobian with
}
\keyword{internal}
