% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_helper.R
\name{specify_hypothesis}
\alias{specify_hypothesis}
\title{(EXPERIMENTAL) This experimental function will soon be deprecated. Please supply a formula or function to the \code{hypothesis} argument to conduct (group-wise) hypothesis tests.}
\usage{
specify_hypothesis(
  hypothesis = "reference",
  comparison = "difference",
  label = NULL,
  label_columns = NULL,
  by = c("term", "group", "contrast"),
  internal = FALSE
)
}
\description{
(EXPERIMENTAL) This experimental function will soon be deprecated. Please supply a formula or function to the \code{hypothesis} argument to conduct (group-wise) hypothesis tests.
}
