% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.marginaleffects}
\alias{summary.marginaleffects}
\title{Summarize a \code{marginaleffects} object}
\usage{
\method{summary}{marginaleffects}(object, ...)
}
\arguments{
\item{object}{An object produced by the \code{marginaleffects} function}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\description{
Summarize a \code{marginaleffects} object
}
