% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.marginalmeans}
\alias{summary.marginalmeans}
\title{Summarize a \code{marginalmeans} object}
\usage{
\method{summary}{marginalmeans}(object, transform_avg = NULL, conf_level = 0.95, ...)
}
\arguments{
\item{object}{An object produced by the \code{marginalmeans} function}

\item{transform_avg}{A function applied to the estimates and confidence intervals \emph{after} the unit-level estimates have been averaged.}

\item{conf_level}{numeric value between 0 and 1. Confidence level to use to build a confidence interval. The default \code{NULL} uses the \code{conf_level} value used in the original call to \code{comparisons()}.}
}
\value{
Data frame of summary statistics for an object produced by the
\code{marginalmeans} function
}
\description{
Summarize a \code{marginalmeans} object
}
\section{Bayesian posterior summaries}{


By default, credible intervals in bayesian models are built as equal-tailed
intervals. This can be changed to a highest density interval by setting a global
option:

\code{options("marginaleffects_posterior_interval" = "eti")}

\code{options("marginaleffects_posterior_interval" = "hdi")}

By default, the center of the posterior distribution in bayesian models is
identified by the median. Users can use a different summary function by setting a
global option:

\code{options("marginaleffects_posterior_center" = mean)}

\code{options("marginaleffects_posterior_center" = median)}

When estimates are averaged using the \code{by} argument, the \code{tidy()} function, or
the \code{summary()} function, the posterior distribution is marginalized twice over.
First, we take the average \emph{across} units but \emph{within} each iteration of the
MCMC chain, according to what the user requested in \code{by} argument or
\code{tidy()/summary()} functions. Then, we identify the center of the resulting
posterior using the function supplied to the
\code{"marginaleffects_posterior_center"} option (the median by default).
}

\seealso{
Other summary: 
\code{\link{aggregate.comparisons}()},
\code{\link{glance.slopes}()},
\code{\link{reexports}},
\code{\link{summary.comparisons}()},
\code{\link{summary.predictions}()},
\code{\link{summary.slopes}()},
\code{\link{tidy.marginalmeans}()}
}
\concept{summary}
