% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_models.R
\docType{data}
\name{supported_models}
\alias{supported_models}
\title{List of model types supported by the \code{marginaleffects} package}
\format{
data.frame
\describe{
\item{Model}{package::function of the model type}
\item{Support: Effect}{Can \code{marginaleffects} retrieve marginal effects?}
\item{Support: Std.Errors}{Can \code{marginaleffects} retrieve standard errors?}
\item{Validty: Stata}{Has the validity of numerical results been checked against Stata's margins command?}
\item{Validty: margins}{Has the validity of numerical results been checked against R's margins package?}
}
}
\usage{
supported_models
}
\description{
A list of model types supported by the \code{marginaleffects} package
}
\keyword{datasets}
