% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{tidy.comparisons}
\alias{tidy.comparisons}
\alias{tidy.slopes}
\alias{tidy.predictions}
\title{Blah}
\usage{
\method{tidy}{comparisons}(x, ...)

\method{tidy}{slopes}(x, ...)

\method{tidy}{predictions}(x, ...)
}
\description{
Blah
}
