% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.hypotheses}
\alias{tidy.hypotheses}
\title{Tidy a \code{hypotheses} object}
\usage{
\method{tidy}{hypotheses}(x, ...)
}
\description{
Tidy a \code{hypotheses} object
}
\seealso{
Other summary: 
\code{\link{glance.slopes}()},
\code{\link{reexports}},
\code{\link{summary.comparisons}()},
\code{\link{summary.marginalmeans}()},
\code{\link{summary.predictions}()},
\code{\link{summary.slopes}()},
\code{\link{tidy.marginalmeans}()}
}
\concept{summary}
