% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.predictions.R
\name{tidy.predictions}
\alias{tidy.predictions}
\title{Tidy a \code{predictions} object}
\usage{
\method{tidy}{predictions}(x, conf_level = NULL, transform_avg = NULL, ...)
}
\arguments{
\item{x}{An object produced by the \code{predictions} function.}

\item{conf_level}{numeric value between 0 and 1. Confidence level to use to build a confidence interval. The default \code{NULL} uses the \code{conf_level} value used in the original call to \code{predictions()}.}

\item{transform_avg}{A function applied to the estimates and confidence intervals \emph{after} the unit-level estimates have been averaged.}
}
\value{
A "tidy" \code{data.frame} of summary statistics which conforms to the
\code{broom} package specification.
}
\description{
Calculate average adjusted predictions by taking the mean of all the
unit-level adjusted predictions computed by the \code{predictions} function.
}
\examples{
mod <- lm(mpg ~ hp * wt + factor(gear), data = mtcars)
mfx <- predictions(mod)
tidy(mfx)
}
\seealso{
Other summary: 
\code{\link{glance.slopes}()},
\code{\link{reexports}},
\code{\link{summary.comparisons}()},
\code{\link{summary.marginalmeans}()},
\code{\link{summary.predictions}()},
\code{\link{summary.slopes}()},
\code{\link{tidy.comparisons}()},
\code{\link{tidy.deltamethod}()},
\code{\link{tidy.marginalmeans}()},
\code{\link{tidy.slopes}()}
}
\concept{summary}
