% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_dictionary.R
\docType{data}
\name{type_dictionary}
\alias{type_dictionary}
\title{type dictionary}
\format{
An object of class \code{data.frame} with 88 rows and 5 columns.
}
\usage{
type_dictionary
}
\description{
insight::get_predict accepts a \code{predict} argument
stats::predict accepts a \code{type} argument
this dictionary converts
}
\keyword{datasets}
