% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_marginaleffects.R
\name{update_marginaleffects}
\alias{update_marginaleffects}
\title{Update \code{marginaleffects} and its dependencies}
\usage{
update_marginaleffects(source = "development")
}
\arguments{
\item{source}{one of two strings: "development" or "cran"}
}
\description{
Update \code{marginaleffects} and its dependencies to the latest R-Universe or CRAN versions. The R session needs to be restarted after install.
}
