% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.comparisons.R
\name{aggregate.comparisons}
\alias{aggregate.comparisons}
\title{This is a test}
\usage{
\method{aggregate}{comparisons}(x, by = NULL, conf_level = NULL, transform_avg = NULL, ...)
}
\description{
This is a test
}
