% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{comparisons}
\alias{comparisons}
\title{Contrasts between adjusted predictions}
\usage{
comparisons(
  model,
  variable,
  newdata = insight::get_data(model),
  type = "response",
  contrast_factor = "reference",
  contrast_numeric = 1,
  contrast_numeric_slope = FALSE,
  ...
)
}
\arguments{
\item{model}{Model object}

\item{newdata}{A dataset over which to compute marginal effects. \code{NULL} uses
the original data used to fit the model.}

\item{type}{Type(s) of prediction as string or character vector. This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{...}{The "Model-Specific Arguments" section below gives a list of arguments which can modify the behavior of this function for certain models (e.g., mixed-effects or bayesian).}
}
\description{
This function calculates contrasts (or comparisons) between adjusted
predictions for each row of the dataset. The resulting object can processed
by the \code{tidy()} or \code{summary()} functions, which compute Average Contrasts.
The \code{datagrid()} function and the \code{newdata} argument can be used to
calculate contrasts Contrasts at the Mean or Contrasts at User-Specified
values (aka Contrasts at Representative values). Additional information can
be found in the Details and Examples sections below, and in the vignette on
the \code{marginaleffects} website.
}
\details{
A "contrast" is the difference between two adjusted predictions, calculated
for meaningfully different regressor values (e.g., College graduates vs.
Others). Uncertainty estimates are computed using the delta method.

Detailed vignettes on contrasts, marginal effects, predictions, and marginal
means, as well as a list of supported models can be found on the package
website:

https://vincentarelbundock.github.io/marginaleffects/
}
