% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{comparisons}
\alias{comparisons}
\title{Contrasts Between Adjusted Predictions}
\usage{
comparisons(
  model,
  newdata = NULL,
  variables = NULL,
  type = "response",
  vcov = TRUE,
  conf.level = 0.95,
  contrast_factor = "reference",
  contrast_numeric = 1,
  ...
)
}
\arguments{
\item{model}{Model object}

\item{newdata}{A dataset over which to compute marginal effects. \code{NULL} uses
the original data used to fit the model.}

\item{variables}{Variables to consider (character vector). \code{NULL}
calculates marginal effects for all terms in the model object.}

\item{type}{Type(s) of prediction as string or character vector. This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{vcov}{Matrix or boolean
\itemize{
\item FALSE: does not compute unit-level standard errors. This can speed up computation considerably.
\item TRUE: computes unit-level standard errors using the default \code{vcov(model)} variance-covariance matrix.
\item Named square matrix: computes standard errors with a user-supplied variance-covariance matrix. This matrix must be square and have dimensions equal to the number of coefficients in \code{get_coef(model)}.
}}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int=TRUE}. Must be strictly greater than 0 and less than 1. Defaults
to 0.95, which corresponds to a 95 percent confidence interval.}

\item{contrast_factor}{string
\itemize{
\item "reference": Each factor level is compared to the factor reference (base) level
\item "sequential": Each factor level is compared to the previous factor level
\item "pairwise": Each factor level is compared to all other levels
}}

\item{contrast_numeric}{string or numeric
\itemize{
\item Numeric of length 1: Contrast between the observed value and the observed value plus \code{contrast_numeric}
\item Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the \code{contrast_numeric} vector.
\item "iqr": Contrast across the interquartile range of the regressor.
\item "sd": Contrast across one standard deviation around the regressor mean.
\item "2sd": Contrast across two standard deviations around the regressor mean.
\item "minmax": Contrast between the maximum and the minimum values of the regressor.
}}

\item{...}{Additional arguments are passed to the \code{predict()} method used to
compute adjusted predictions. These arguments are particularly useful for
mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?marginaleffects} documentation for a non-exhaustive list of available
arguments.}
}
\description{
This function calculates contrasts (or comparisons) between adjusted
predictions for each row of the dataset. The resulting object can processed
by the \code{tidy()} or \code{summary()} functions, which compute Average Contrasts.
The \code{datagrid()} function and the \code{newdata} argument can be used to
calculate contrasts Contrasts at the Mean or Contrasts at User-Specified
values (aka Contrasts at Representative values). Additional information can
be found in the Details and Examples sections below, and in the vignette on
the \code{marginaleffects} website.
}
\details{
A "contrast" is the difference between two adjusted predictions, calculated
for meaningfully different regressor values (e.g., College graduates vs.
Others). Uncertainty estimates are computed using the delta method.

Detailed vignettes on contrasts, marginal effects, predictions, and marginal
means, as well as a list of supported models can be found on the package
website:

https://vincentarelbundock.github.io/marginaleffects/
}
\section{Model-Specific Arguments}{


Some model types allow model-specific arguments to modify the nature of
marginal effects, predictions, marginal means, and contrasts.\tabular{llll}{
   Package \tab Class \tab Argument \tab Documentation \cr
   \code{brms} \tab \code{brmsfit} \tab \code{ndraws} \tab \link[brms:posterior_predict.brmsfit]{brms::posterior_predict} \cr
    \tab  \tab \code{re_formula} \tab  \cr
   \code{lme4} \tab \code{merMod} \tab \code{include_random} \tab \link[insight:get_predicted]{insight::get_predicted} \cr
    \tab  \tab \code{re.form} \tab \link[lme4:predict.merMod]{lme4::predict.merMod} \cr
    \tab  \tab \code{allow.new.levels} \tab \link[lme4:predict.merMod]{lme4::predict.merMod} \cr
   \code{glmmTMB} \tab \code{glmmTMB} \tab \code{re.form} \tab \link[glmmTMB:predict.glmmTMB]{glmmTMB::predict.glmmTMB} \cr
   \code{glmmTMB} \tab \code{glmmTMB} \tab \code{allow.new.levels} \tab \link[glmmTMB:predict.glmmTMB]{glmmTMB::predict.glmmTMB} \cr
   \code{glmmTMB} \tab \code{glmmTMB} \tab \code{zitype} \tab \link[glmmTMB:predict.glmmTMB]{glmmTMB::predict.glmmTMB} \cr
   \code{mgcv} \tab \code{bam} \tab \code{exclude} \tab \link[mgcv:predict.bam]{mgcv::predict.bam} \cr
   \code{robustlmm} \tab \code{rlmerMod} \tab \code{re.form} \tab \link[robustlmm:rlmerMod-class]{robustlmm::predict.rlmerMod} \cr
   \code{robustlmm} \tab \code{rlmerMod} \tab \code{allow.new.levels} \tab \link[robustlmm:rlmerMod-class]{robustlmm::predict.rlmerMod} \cr
}
}

\section{Parallel computation}{


The \code{marginaleffects()} and \code{comparisons()} functions can use parallelism to
speed up computation. Operations are parallelized at the "term" or
"variable" level. This means that no speed gain is available when the user
computes marginal effects for only one right-hand side variable. There is
always some overhead when using parallel computation. Thus, parallel
computation is most likely to be useful when:
\enumerate{
\item The model includes many right-hand side variables.
\item \code{marginaleffects} must compute pairwise contrasts for multiple factor variables.
}

To activate parallel computation, users must load the \code{future.apply} package
and call \code{plan()} function. For example:\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(future.apply)
plan("multisession")
marginaleffects(model)
}\if{html}{\out{</div>}}

To disable parallelism in \code{marginaleffects} altogether, you can set a global option:\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(marginaleffects_parallel = FALSE)
}\if{html}{\out{</div>}}
}

\examples{

library(marginaleffects)
library(magrittr)

# Linear model
tmp <- mtcars
tmp$am <- as.logical(tmp$am)
mod <- lm(mpg ~ am + factor(cyl), tmp)
comparisons(mod, contrast_factor = "reference") \%>\% tidy()
comparisons(mod, contrast_factor = "sequential") \%>\% tidy()
comparisons(mod, contrast_factor = "pairwise") \%>\% tidy()

# GLM with different scale types
mod <- glm(am ~ factor(gear), data = mtcars)
comparisons(mod, type = "response") \%>\% tidy()
comparisons(mod, type = "link") \%>\% tidy()

# Numeric contrasts
mod <- lm(mpg ~ hp, data = mtcars)
comparisons(mod, contrast_numeric = 1) \%>\% tidy()
comparisons(mod, contrast_numeric = 5) \%>\% tidy()
comparisons(mod, contrast_numeric = c(90, 100)) \%>\% tidy()
comparisons(mod, contrast_numeric = "iqr") \%>\% tidy()
comparisons(mod, contrast_numeric = "sd") \%>\% tidy()
comparisons(mod, contrast_numeric = "minmax") \%>\% tidy()

}
