% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual.R
\name{counterfactual}
\alias{counterfactual}
\title{Generate "counterfactual" datasets based on a model}
\usage{
counterfactual(model = NULL, data = NULL, at = NULL)
}
\description{
Generate "counterfactual" datasets based on a model
}
