% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{counterfactual}
\alias{counterfactual}
\title{Superseded by \code{datagridcf}}
\usage{
counterfactual(..., model = NULL, newdata = NULL)
}
\description{
Superseded by \code{datagridcf}
}
\keyword{internal}
