% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotheses.R
\name{deltamethod}
\alias{deltamethod}
\title{\code{deltamethod()} is an alias to \code{hypotheses()}}
\usage{
deltamethod(
  model,
  hypothesis = NULL,
  FUN = NULL,
  vcov = NULL,
  conf_level = 0.95,
  null = 0,
  delta = 0,
  side = NULL,
  df = Inf,
  ...
)
}
\description{
This alias is kept for backward compatibility.
}
\keyword{internal}
