% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotheses.R
\name{deltamethod}
\alias{deltamethod}
\title{\code{deltamethod()} is an alias to \code{hypotheses()}. It is kept for backward compatibility.}
\usage{
deltamethod(
  model,
  hypothesis = NULL,
  FUN = NULL,
  vcov = NULL,
  conf_level = 0.95,
  ...
)
}
\description{
This alias is kept for backward compatibility and because some users may prefer that name.
}
\keyword{internal}
