% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotheses.R
\name{deltamethod}
\alias{deltamethod}
\title{\code{deltamethod()} is an alias to \code{hypotheses()}}
\usage{
deltamethod(
  model,
  hypothesis = NULL,
  vcov = NULL,
  conf_level = 0.95,
  df = Inf,
  equivalence = NULL,
  joint = FALSE,
  joint_test = "f",
  FUN = NULL,
  numderiv = "fdforward",
  ...
)
}
\description{
This alias is kept for backward compatibility.
}
\keyword{internal}
