% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R, R/methods_betareg.R
\name{get_coef}
\alias{get_coef}
\alias{get_coef.betareg}
\title{Get a named vector of coefficients}
\usage{
get_coef(model, ...)

\method{get_coef}{betareg}(model)
}
\description{
Get a named vector of coefficients
}
