% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R, R/methods_MASS.R,
%   R/methods_betareg.R
\name{get_coef}
\alias{get_coef}
\alias{get_coef.default}
\alias{get_coef.polr}
\alias{get_coef.betareg}
\title{Get a named vector of coefficients from a model object (internal function)}
\usage{
get_coef(model, ...)

\method{get_coef}{default}(model, ...)

\method{get_coef}{polr}(model, ...)

\method{get_coef}{betareg}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
A named vector of coefficients. The names must match those of the variance matrix.
}
\description{
Get a named vector of coefficients from a model object (internal function)
}
\keyword{internal}
