% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R, R/methods_MASS.R,
%   R/methods_betareg.R, R/methods_brglm2.R, R/methods_brms.R, R/methods_lme4.R,
%   R/methods_mgcv.R, R/methods_scam.R
\name{get_coef}
\alias{get_coef}
\alias{get_coef.default}
\alias{get_coef.polr}
\alias{get_coef.betareg}
\alias{get_coef.brmultinom}
\alias{get_coef.bracl}
\alias{get_coef.brmsfit}
\alias{get_coef.lmerModLmerTest}
\alias{get_coef.lmerMod}
\alias{get_coef.gam}
\alias{get_coef.scam}
\title{Get a named vector of coefficients from a model object (internal function)}
\usage{
get_coef(model, ...)

\method{get_coef}{default}(model, ...)

\method{get_coef}{polr}(model, ...)

\method{get_coef}{betareg}(model, ...)

\method{get_coef}{brmultinom}(model, ...)

\method{get_coef}{bracl}(model, ...)

\method{get_coef}{brmsfit}(model, ...)

\method{get_coef}{lmerModLmerTest}(model, ...)

\method{get_coef}{lmerMod}(model, ...)

\method{get_coef}{gam}(model, ...)

\method{get_coef}{scam}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{The "Model-Specific Arguments" section below gives a list of arguments which can modify the behavior of this function for certain models (e.g., mixed-effects or bayesian).}
}
\value{
A named vector of coefficients. The names must match those of the variance matrix.
}
\description{
Get a named vector of coefficients from a model object (internal function)
}
\keyword{internal}
