% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R, R/methods_MASS.R,
%   R/methods_betareg.R, R/methods_brglm2.R, R/methods_brms.R, R/methods_lme4.R,
%   R/methods_mgcv.R, R/methods_sampleSelection.R, R/methods_scam.R
\name{get_coef}
\alias{get_coef}
\alias{get_coef.default}
\alias{get_coef.polr}
\alias{get_coef.betareg}
\alias{get_coef.brmultinom}
\alias{get_coef.bracl}
\alias{get_coef.brmsfit}
\alias{get_coef.lmerModLmerTest}
\alias{get_coef.lmerMod}
\alias{get_coef.gam}
\alias{get_coef.selection}
\alias{get_coef.scam}
\title{Get a named vector of coefficients from a model object (internal function)}
\usage{
get_coef(model, ...)

\method{get_coef}{default}(model, ...)

\method{get_coef}{polr}(model, ...)

\method{get_coef}{betareg}(model, ...)

\method{get_coef}{brmultinom}(model, ...)

\method{get_coef}{bracl}(model, ...)

\method{get_coef}{brmsfit}(model, ...)

\method{get_coef}{lmerModLmerTest}(model, ...)

\method{get_coef}{lmerMod}(model, ...)

\method{get_coef}{gam}(model, ...)

\method{get_coef}{selection}(model, ...)

\method{get_coef}{scam}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are passed to the \code{predict()} method used to
compute adjusted predictions. These arguments are particularly useful for
mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?marginaleffects} document for a non-exhaustive list of available
arguments.}
}
\value{
A named vector of coefficients. The names must match those of the variance matrix.
}
\description{
Get a named vector of coefficients from a model object (internal function)
}
\keyword{internal}
