% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrast.R
\name{get_contrast}
\alias{get_contrast}
\title{Get contrasts for factor and logical terms in a model object (internal function)}
\usage{
get_contrast(model, variable, prediction_type = "response", ...)
}
\arguments{
\item{model}{Model object}

\item{variable}{Name of the variable to contrast (string)}

\item{prediction_type}{Type(s) of prediction as string or vector This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{...}{Additional arguments are ignored.}
}
\value{
A \code{broom}-style \code{data.frame} with the results of the
\code{emmeans::contrast} function.
}
\description{
Get contrasts for factor and logical terms in a model object (internal function)
}
\keyword{internal}
