% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Download and Read Dataset}
\usage{
get_dataset(dataset = "thornton", package = "marginaleffects", docs = FALSE)
}
\arguments{
\item{dataset}{A string, either "military" or "thornton".}

\item{docs}{A logical, if TRUE open the documentation using \code{getOption("viewer")}}
}
\value{
A data frame containing the dataset.
}
\description{
Downloads a dataset and reads it into a data frame.
}
