% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R, R/get_dydx_stats.R,
%   R/get_dydx_nnet.R, R/get_dydx_MASS.R, R/get_dydx_betareg.R,
%   R/get_dydx_fixest.R, R/get_dydx_ivreg.R, R/get_dydx_lme4.R,
%   R/get_dydx_ordinal.R, R/get_dydx_survey.R
\name{get_dydx}
\alias{get_dydx}
\alias{get_dydx.glm}
\alias{get_dydx.loess}
\alias{get_dydx.multinom}
\alias{get_dydx.polr}
\alias{get_dydx.betareg}
\alias{get_dydx.fixest}
\alias{get_dydx.ivreg}
\alias{get_dydx.glmerMod}
\alias{get_dydx.lmerMod}
\alias{get_dydx.clm}
\alias{get_dydx.svyglm}
\title{Get dY/dX}
\usage{
get_dydx(model, ...)

\method{get_dydx}{glm}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{loess}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{multinom}(
  model,
  fitfram,
  variable,
  group_name,
  variance,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{polr}(
  model,
  fitfram,
  variable,
  group_name,
  variance,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{fixest}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{ivreg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{glmerMod}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{lmerMod}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{clm}(
  model,
  fitfram,
  variable,
  variance = NULL,
  prediction_type = "prob",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{svyglm}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get dY/dX
}
