% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydx.R
\name{get_dydx}
\alias{get_dydx}
\alias{get_dydx.default}
\title{Get Gradient}
\usage{
get_dydx(model, ...)

\method{get_dydx}{default}(
  model,
  variable,
  fitfram = insight::get_data(model),
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get Gradient
}
