% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydx_nnet.R, R/get_gradient.R
\name{get_gradient.multinom}
\alias{get_gradient.multinom}
\alias{get_gradient}
\alias{get_gradient.default}
\title{Get Gradient}
\usage{
\method{get_gradient}{multinom}(
  model,
  fitfram,
  variable,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple"
)

get_gradient(model, ...)

\method{get_gradient}{default}(
  model,
  fitfram,
  variable,
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get Gradient
}
