% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_names.R, R/methods_MASS.R,
%   R/methods_nnet.R, R/methods_ordinal.R
\name{get_group_names}
\alias{get_group_names}
\alias{get_group_names.default}
\alias{get_group_names.polr}
\alias{get_group_names.multinom}
\alias{get_group_names.clm}
\title{Get levels of the outcome variable in grouped or multivariate models}
\usage{
get_group_names(model, ...)

\method{get_group_names}{default}(model, ...)

\method{get_group_names}{polr}(model, ...)

\method{get_group_names}{multinom}(model, ...)

\method{get_group_names}{clm}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
A character vector
}
\description{
Get levels of the outcome variable in grouped or multivariate models
}
\keyword{internal}
