% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R, R/get_dydx_stats.R,
%   R/get_dydx_nnet.R, R/get_dydx_MASS.R, R/get_dydx_betareg.R,
%   R/get_dydx_fixest.R, R/get_dydx_ivreg.R, R/get_dydx_lme4.R,
%   R/get_dydx_ordinal.R, R/get_dydx_survey.R
\name{get_jacobian}
\alias{get_jacobian}
\alias{get_jacobian.glm}
\alias{get_jacobian.lm}
\alias{get_dydx.lm}
\alias{get_jacobian.multinom}
\alias{get_jacobian.polr}
\alias{get_jacobian.betareg}
\alias{get_jacobian.fixest}
\alias{get_jacobian.ivreg}
\alias{get_jacobian.glmerMod}
\alias{get_jacobian.lmerMod}
\alias{get_jacobian.clm}
\alias{get_jacobian.svyglm}
\title{Get Jacobian}
\usage{
get_jacobian(model, ...)

\method{get_jacobian}{glm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{lm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_dydx}{lm}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{multinom}(
  model,
  fitfram,
  variable,
  variance = NULL,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{polr}(
  model,
  fitfram,
  variable,
  variance = NULL,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{fixest}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{ivreg}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{glmerMod}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{lmerMod}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{clm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_jacobian}{svyglm}(
  model,
  fitfram,
  variable,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get Jacobian
}
