% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalmeans.R
\name{get_marginalmeans}
\alias{get_marginalmeans}
\title{Workhorse function for \code{marginalmeans}}
\usage{
get_marginalmeans(model, newdata, type, variables, ...)
}
\arguments{
\item{model}{Model object}

\item{newdata}{A dataset over which to compute adjusted predictions. \code{NULL} uses
the original data used to fit the model.}

\item{type}{Type(s) of prediction (string or character vector). This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{variables}{Categorical predictors over which to compute marginal means
(character vector). \code{NULL} calculates marginal means for all logical,
character, or factor variables in the dataset used to fit \code{model}.}

\item{...}{absorb useless arguments from other get_* workhorse functions}
}
\description{
Needs to be separate because we also need it in \code{delta_method}
}
