% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gradient.R, R/methods_nnet.R
\name{get_mfx}
\alias{get_mfx}
\alias{get_mfx.default}
\alias{get_mfx.multinom}
\title{Get Gradient}
\usage{
get_mfx(model, ...)

\method{get_mfx}{default}(
  model,
  fitfram,
  variable,
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_mfx}{multinom}(
  model,
  fitfram,
  variable,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple"
)
}
\description{
Get Gradient
}
