% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dydx.R, R/methods_betareg.R,
%   R/methods_nnet.R
\name{get_mfx_and_se}
\alias{get_mfx_and_se}
\alias{get_mfx_and_se.default}
\alias{get_mfx_and_se.betareg}
\alias{get_mfx_and_se.multinom}
\title{Get dY/dX}
\usage{
get_mfx_and_se(model, ...)

\method{get_mfx_and_se}{default}(
  model,
  fitfram,
  variable,
  variance,
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_mfx_and_se}{betareg}(
  model,
  fitfram,
  variable,
  variance,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_mfx_and_se}{multinom}(
  model,
  fitfram,
  variable,
  group_name,
  variance,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)
}
\description{
Get dY/dX
}
