% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict.R, R/methods_ordinal.R
\name{get_predict}
\alias{get_predict}
\alias{get_predict.clm}
\title{Get a named vector of predictficients from a model object}
\usage{
get_predict(model, newdata, prediction_type, ...)

\method{get_predict}{clm}(model, newdata, prediction_type, ...)
}
\description{
Get a named vector of predictficients from a model object
}
