% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predict.R, R/methods_MASS.R,
%   R/methods_afex.R, R/methods_aod.R, R/methods_betareg.R, R/methods_bife.R,
%   R/methods_biglm.R, R/methods_nnet.R, R/methods_brglm2.R, R/methods_brms.R,
%   R/methods_crch.R, R/methods_fixest.R, R/methods_gamlss.R,
%   R/methods_glmmTMB.R, R/methods_lme4.R, R/methods_mclogit.R,
%   R/methods_mhurdle.R, R/methods_mlogit.R, R/methods_ordinal.R,
%   R/methods_quantreg.R, R/methods_robustlmm.R, R/methods_rstanarm.R,
%   R/methods_survival.R, R/methods_tobit1.R
\name{get_predict}
\alias{get_predict}
\alias{get_predict.default}
\alias{get_predict.polr}
\alias{get_predict.glmmPQL}
\alias{get_predict.afex_aov}
\alias{get_predict.glimML}
\alias{get_predict.betareg}
\alias{get_predict.bife}
\alias{get_predict.biglm}
\alias{get_predict.multinom}
\alias{get_predict.brmultinom}
\alias{get_predict.brmsfit}
\alias{get_predict.crch}
\alias{get_predict.fixest}
\alias{get_predict.gamlss}
\alias{get_predict.glmmTMB}
\alias{get_predict.merMod}
\alias{get_predict.lmerModLmerTest}
\alias{get_predict.lmerMod}
\alias{get_predict.mblogit}
\alias{get_predict.mhurdle}
\alias{get_predict.mlogit}
\alias{get_predict.clm}
\alias{get_predict.rq}
\alias{get_predict.rlmerMod}
\alias{get_predict.stanreg}
\alias{get_predict.coxph}
\alias{get_predict.tobit1}
\title{Get predicted values from a model object (internal function)}
\usage{
get_predict(model, newdata, vcov, conf_level, type, ...)

\method{get_predict}{default}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{polr}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "probs",
  ...
)

\method{get_predict}{glmmPQL}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{afex_aov}(model, newdata = NULL, ...)

\method{get_predict}{glimML}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{betareg}(model, newdata, ...)

\method{get_predict}{bife}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{biglm}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{multinom}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "probs",
  ...
)

\method{get_predict}{brmultinom}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "probs",
  ...
)

\method{get_predict}{brmsfit}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{crch}(
  model,
  newdata = NULL,
  vcov = FALSE,
  conf_level = 0.95,
  type = "location",
  ...
)

\method{get_predict}{fixest}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{gamlss}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{glmmTMB}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{merMod}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{lmerModLmerTest}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{lmerMod}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{mblogit}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{mhurdle}(
  model,
  newdata = insight::get_data(model),
  vcov = NULL,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{mlogit}(model, newdata, ...)

\method{get_predict}{clm}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{rq}(
  model,
  newdata = insight::get_data(model),
  vcov = NULL,
  conf_level = 0.95,
  type = NULL,
  ...
)

\method{get_predict}{rlmerMod}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{stanreg}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "response",
  ...
)

\method{get_predict}{coxph}(
  model,
  newdata = insight::get_data(model),
  vcov = FALSE,
  conf_level = 0.95,
  type = "lp",
  ...
)

\method{get_predict}{tobit1}(
  model,
  newdata = insight::get_data(model),
  vcov = NULL,
  conf_level = 0.95,
  type = "response",
  ...
)
}
\value{
A data.frame of predicted values with a number of rows equal to the
number of rows in \code{newdata} and columns "rowid" and "predicted". A "group"
column is added for multivariate models or models with categorical outcomes.
}
\description{
Get predicted values from a model object (internal function)
}
\keyword{internal}
