% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_delta.R, R/methods_nnet.R
\name{get_se_delta}
\alias{get_se_delta}
\alias{get_se_delta.default}
\alias{get_se_delta.multinom}
\title{Compute a vector of standard errors}
\usage{
get_se_delta(model, ...)

\method{get_se_delta}{default}(
  model,
  variable,
  fitfram = insight::find_data(model),
  variance = vcov(model),
  group_name = NULL,
  prediction_type = "response",
  numDeriv_method = "simple",
  ...
)

\method{get_se_delta}{multinom}(
  model,
  variable,
  fitfram = insight::find_data(model),
  variance = NULL,
  group_name = NULL,
  prediction_type = "probs",
  numDeriv_method = "simple",
  ...
)
}
\description{
Compute a vector of standard errors
}
