% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcov.R, R/methods_afex.R,
%   R/methods_aod.R, R/methods_biglm.R, R/methods_brms.R, R/methods_gamlss.R,
%   R/methods_mhurdle.R, R/methods_rms.R, R/methods_scam.R
\name{get_vcov}
\alias{get_vcov}
\alias{get_vcov.default}
\alias{get_vcov.afex_aov}
\alias{get_vcov.glimML}
\alias{get_vcov.biglm}
\alias{get_vcov.brmsfit}
\alias{get_vcov.gamlss}
\alias{get_vcov.mhurdle}
\alias{get_vcov.orm}
\alias{get_vcov.scam}
\title{Get a named variance-covariance matrix from a model object (internal function)}
\usage{
get_vcov(model, ...)

\method{get_vcov}{default}(model, vcov = NULL, ...)

\method{get_vcov}{afex_aov}(model, vcov = NULL, ...)

\method{get_vcov}{glimML}(model, vcov = NULL, ...)

\method{get_vcov}{biglm}(model, vcov = NULL, ...)

\method{get_vcov}{brmsfit}(model, vcov = NULL, ...)

\method{get_vcov}{gamlss}(model, ...)

\method{get_vcov}{mhurdle}(model, vcov = NULL, ...)

\method{get_vcov}{orm}(model, vcov = NULL, ...)

\method{get_vcov}{scam}(model, vcov = NULL, ...)
}
\value{
A named square matrix of variance and covariances. The names must match the coefficient names.
}
\description{
Get a named variance-covariance matrix from a model object (internal function)
}
\keyword{internal}
