% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcov.R, R/methods_MASS.R, R/methods_crch.R,
%   R/methods_nnet.R
\name{get_vcov}
\alias{get_vcov}
\alias{get_vcov.default}
\alias{get_vcov.polr}
\alias{get_vcov.crch}
\alias{get_vcov.multinom}
\title{Get a named variance-covariance matrix from a model object (internal function)}
\usage{
get_vcov(model, ...)

\method{get_vcov}{default}(model, ...)

\method{get_vcov}{polr}(model, ...)

\method{get_vcov}{crch}(model, ...)

\method{get_vcov}{multinom}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
A named square matrix of variance and covariances. The names must match the coefficient names.
}
\description{
Get a named variance-covariance matrix from a model object (internal function)
}
\keyword{internal}
