% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcov.R, R/methods_MASS.R, R/methods_aod.R,
%   R/methods_nnet.R, R/methods_crch.R, R/methods_scam.R
\name{get_vcov}
\alias{get_vcov}
\alias{get_vcov.default}
\alias{get_vcov.polr}
\alias{get_vcov.glimML}
\alias{get_vcov.multinom}
\alias{get_vcov.crch}
\alias{get_vcov.scam}
\title{Get a named variance-covariance matrix from a model object (internal function)}
\usage{
get_vcov(model, ...)

\method{get_vcov}{default}(model, ...)

\method{get_vcov}{polr}(model, ...)

\method{get_vcov}{glimML}(model, ...)

\method{get_vcov}{multinom}(model, ...)

\method{get_vcov}{crch}(model, ...)

\method{get_vcov}{scam}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{The "Model-Specific Arguments" section below gives a list of arguments which can modify the behavior of this function for certain models (e.g., mixed-effects or bayesian).}
}
\value{
A named square matrix of variance and covariances. The names must match the coefficient names.
}
\description{
Get a named variance-covariance matrix from a model object (internal function)
}
\keyword{internal}
