% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.marginaleffects}
\alias{glance.marginaleffects}
\title{Glance at key characteristics of an object}
\usage{
\method{glance}{marginaleffects}(x, ...)
}
\arguments{
\item{x}{An object produced by the \code{marginaleffects} function.}

\item{...}{Additional arguments are passed to the \code{predict()} method
supplied by the modeling package.These arguments are particularly useful
for mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?marginaleffects} documentation for a non-exhaustive list of available
arguments.}
}
\description{
Glance at key characteristics of an object
}
