% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.slopes}
\alias{glance.slopes}
\title{Glance at key characteristics of an object}
\usage{
\method{glance}{slopes}(x, ...)
}
\description{
Glance at key characteristics of an object
}
\seealso{
Other summary: 
\code{\link{reexports}},
\code{\link{summary.comparisons}()},
\code{\link{summary.marginalmeans}()},
\code{\link{summary.predictions}()},
\code{\link{summary.slopes}()},
\code{\link{tidy.hypotheses}()},
\code{\link{tidy.marginalmeans}()}
}
\concept{summary}
