% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferences.R
\name{inferences}
\alias{inferences}
\title{EXPERIMENTAL Uncertainty Estimates for \code{marginaleffects} Objects}
\usage{
inferences(x, method = "simulation", R = 1000, conf_type = "perc", ...)
}
\arguments{
\item{x}{Object produced by one of the core \code{marginaleffects} functions.}

\item{method}{String
\itemize{
\item "delta": delta method standard errors
\item "boot" package
\item "rsample" package
\item "simulation" from a multivariate normal distribution (Krinsky & Robb, 1986)
}}

\item{R}{Number of resamples or simulations.}

\item{conf_type}{String: type of bootstrap interval to construct.
\itemize{
\item \code{boot}: "perc", "norm", "basic", or "bca"
\item \code{rsample}: "perc" or "bca"
\item \code{simulation}: argument ignored.
}}

\item{...}{\itemize{
\item If \code{method="boot"}, additional arguments are passed to \code{boot::boot()}.
\item If \code{method="rsample"}, additional arguments are passed to \code{rsample::bootstraps()}.
\item If \code{method="simulation"}, additional arguments are ignored.
}}
}
\value{
A \code{marginaleffects} object with simulation or bootstrap resamples and objects attached.
}
\description{
Warning: This function is experimental. It may be renamed, the user interface may change, or the functionality may migrate to arguments in other \code{marginaleffects} functions.

Apply this function to a \code{marginaleffects} object to change the inferential method used to compute uncertainty estimates.
}
\details{
When \code{method="simulation"}, we conduct simulation-based inference following the method discussed in Krinsky & Robb (1986):
\enumerate{
\item Draw \code{R} sets of simulated coefficients from a multivariate normal distribution with mean equal to the original model's estimated coefficients and variance equal to the model's variance-covariance matrix (classical, "HC3", or other).
\item Use the \code{R} sets of coefficients to compute \code{R} sets of estimands: predictions, comparisons, or slopes.
\item Take quantiles of the resulting distribution of estimands to obtain a confidence interval and the standard deviation of simulated estimates to estimate the standard error.
}
}
\section{References}{


Krinsky, I., and A. L. Robb. 1986. “On Approximating the Statistical Properties of Elasticities.” Review of Economics and Statistics 68 (4): 715–9.

King, Gary, Michael Tomz, and Jason Wittenberg. "Making the most of statistical analyses: Improving interpretation and presentation." American journal of political science (2000): 347-361

Dowd, Bryan E., William H. Greene, and Edward C. Norton. "Computation of standard errors." Health services research 49.2 (2014): 731-750.
}

\examples{
library(marginaleffects)
library(magrittr)
mod <- glm(vs ~ hp * wt + factor(gear), data = mtcars, family = binomial)

avg_predictions(mod, by = "gear") \%>\%
  inferences()

slopes(mod) \%>\%
  inferences() \%>\%
  head()

avg_slopes(mod) \%>\%
  inferences() \%>\%
  posterior_draws("rvar")

}
