% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects.R
\name{marginaleffects}
\alias{marginaleffects}
\title{compute marginal effects estimates using numerical derivatives}
\usage{
marginaleffects(
  model,
  newdata = NULL,
  variables = NULL,
  variance = TRUE,
  numDeriv_method = "simple",
  prediction_type = "response",
  return_data = TRUE
)
}
\description{
compute marginal effects estimates using numerical derivatives
}
