% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{marginaleffects}
\alias{marginaleffects}
\title{\code{marginaleffects()} is an alias to \code{slopes()}. It is kept for backward compatibility.}
\usage{
marginaleffects(
  model,
  newdata = NULL,
  variables = NULL,
  vcov = TRUE,
  conf_level = 0.95,
  type = NULL,
  slope = "dydx",
  by = NULL,
  wts = NULL,
  hypothesis = NULL,
  eps = NULL,
  ...
)
}
\description{
\code{marginaleffects()} is an alias to \code{slopes()}. It is kept for backward compatibility.
}
\keyword{internal}
