% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects_internal.R
\docType{class}
\name{marginaleffects_internal-class}
\alias{marginaleffects_internal-class}
\title{Internal S4 class for marginaleffects}
\description{
This S4 class is used internally to hold common arguments passed between
functions to simplify the function signatures and reduce redundant argument passing.
}
\section{Slots}{

\describe{
\item{\code{model}}{The fitted model object}

\item{\code{modeldata}}{The model data frame}

\item{\code{newdata}}{The new data frame for predictions}
}}

\keyword{internal}
