% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalmeans.R
\name{marginalmeans}
\alias{marginalmeans}
\title{Marginal means}
\usage{
marginalmeans(
  model,
  newdata = NULL,
  variables = NULL,
  type = "response",
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{model}{Model object}

\item{newdata}{A dataset over which to compute marginal effects. \code{NULL} uses
the original data used to fit the model.}

\item{variables}{Character vector. Compute Estimated Marginal Means for
combinations of each of these variables. Factor levels are considered at
each of their levels. Numeric variables variables are considered at Tukey's
Five-Number Summaries.}

\item{type}{Type(s) of prediction as string or vector This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{conf.level}{The confidence level to use for the confidence interval.
No interval is computed if \code{conf.int=NULL}.  Must be strictly greater than 0
and less than 1. Defaults to 0.95, which corresponds to a 95 percent
confidence interval.}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\description{
Warning: This package is experimental.
}
