% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R
\name{median_or_mode}
\alias{median_or_mode}
\alias{median_or_mode.default}
\alias{median_or_mode.factor}
\alias{median_or_mode.logical}
\alias{median_or_mode.data.frame}
\title{Compute the mode or median of \code{x}}
\usage{
median_or_mode(x)

\method{median_or_mode}{default}(x)

\method{median_or_mode}{factor}(x)

\method{median_or_mode}{logical}(x)

\method{median_or_mode}{data.frame}(x)
}
\arguments{
\item{x}{extract the median or the mode of vector or data.frame x depending on its type}
}
\description{
Compute the mode or median of \code{x}
}
\keyword{internal}
