% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meffects.R
\name{meffects}
\alias{meffects}
\title{compute marginal effects estimates using numerical derivatives}
\usage{
meffects(
  model,
  newdata = NULL,
  at = NULL,
  variables = NULL,
  variance = try(stats::vcov(model), silent = TRUE),
  prediction_type = "response",
  numDeriv_method = "simple"
)
}
\description{
compute marginal effects estimates using numerical derivatives
}
