% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects.R, R/marginalmeans.R
\name{meffects}
\alias{meffects}
\title{\code{meffects()} is a shortcut to \code{marginaleffects()}}
\usage{
meffects(
  model,
  newdata = NULL,
  variables = NULL,
  vcov = TRUE,
  predict_type = "response",
  return_data = TRUE,
  ...
)

meffects(
  model,
  newdata = NULL,
  variables = NULL,
  vcov = TRUE,
  predict_type = "response",
  return_data = TRUE,
  ...
)
}
\arguments{
\item{model}{Model object}

\item{newdata}{A dataset over which to compute marginal effects. \code{NULL} uses
the original data used to fit the model.}

\item{variables}{Variables to consider (character vector). \code{NULL}
calculates marginal effects for all terms in the model object.}

\item{vcov}{Matrix or boolean
\itemize{
\item FALSE: does not compute unit-level standard errors.
\item TRUE: computes unit-level standard errors using the default \code{vcov(model)} variance-covariance matrix.
\item Named square matrix: computes standard errors with a user-supplied variance-covariance matrix. This matrix must be square and have dimensions equal to the number of coefficients in \code{get_coef(model)}.
}}

\item{predict_type}{Type(s) of prediction as string or vector This can
differ based on the model type, but will typically be a string such as:
"response", "link", "probs", or "zero".}

\item{return_data}{boolean If \code{TRUE}, the original data used to fit the
model is attached to the output. \code{FALSE} will objects which take up less
space in memory.}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\description{
Warning: This package is experimental.  A "marginal effect" is the partial
derivative of the regression equation with respect to a variable in the
model. This package uses automatic differentiation tu compute marginal
effects for a vast array of models, including non-linear models with
transformations (e.g., polynomials). The list of supported model types is
available in the \code{supported_models} dataset which accompanies this package:
\code{marginaleffects::supported_models}
}
\examples{

mod <- glm(am ~ hp * wt, data = mtcars, family = binomial)
mfx <- marginaleffects(mod)
summary(mfx)
tidy(mfx)
head(mfx)
\dontrun{
plot(mfx)
}

# typical marginal effects
marginaleffects(mod, 
                newdata = typical(hp = c(100, 110)))

# counterfactual average marginal effects
marginaleffects(mod, 
                newdata = counterfactual(hp = c(100, 110)))

# heteroskedasticity robust standard errors
marginaleffects(mod, vcov = sandwich::vcovHC(mod))
               
}
\keyword{internal}
