% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{meffects}
\alias{meffects}
\title{\code{meffects()} is an alias to \code{slopes()}}
\usage{
meffects(...)
}
\description{
This alias is kept for backward compatibility and because some users may prefer that name.
}
\keyword{internal}
