% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfx.R
\name{mfx}
\alias{mfx}
\title{compute marginal effects estimates using numerical derivatives}
\usage{
mfx(
  model,
  newdata = NULL,
  variables = NULL,
  variance = try(stats::vcov(model), silent = TRUE),
  prediction_type = "response",
  numDeriv_method = "simple"
)
}
\description{
compute marginal effects estimates using numerical derivatives
}
