% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_marginaleffects_internal}
\alias{new_marginaleffects_internal}
\title{Constructor for marginaleffects_internal class}
\usage{
new_marginaleffects_internal(
  model,
  modeldata = data.frame(),
  newdata = data.frame()
)
}
\arguments{
\item{model}{The fitted model object}

\item{modeldata}{The model data frame}

\item{newdata}{The new data frame for predictions}
}
\value{
An object of class marginaleffects_internal
}
\description{
Constructor for marginaleffects_internal class
}
\keyword{internal}
