% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_marginaleffects_internal}
\alias{new_marginaleffects_internal}
\title{Constructor for marginaleffects_internal class}
\usage{
new_marginaleffects_internal(
  model = NULL,
  call = NULL,
  conf_level = 0.95,
  df = NULL,
  draws = NULL,
  hypothesis = NULL,
  hypothesis_null = NULL,
  hypothesis_direction = NULL,
  jacobian = NULL,
  modeldata = data.frame(),
  newdata = data.frame(),
  type = NULL,
  variables = list(),
  vcov_model = NULL,
  wts = NULL
)
}
\arguments{
\item{model}{The fitted model object}

\item{call}{The original function call}

\item{df}{The degrees of freedom}

\item{modeldata}{The model data frame}

\item{newdata}{The new data frame for predictions}

\item{type}{The sanitized type from sanitize_type()}

\item{vcov_model}{The variance-covariance matrix}

\item{wts}{The weights specification}
}
\value{
An object of class marginaleffects_internal
}
\description{
Constructor for marginaleffects_internal class
}
\keyword{internal}
