% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_marginaleffects_internal}
\alias{new_marginaleffects_internal}
\title{Constructor for marginaleffects_internal class}
\usage{
new_marginaleffects_internal(
  model,
  call,
  by = FALSE,
  byfun = NULL,
  comparison = NULL,
  conf_level = 0.95,
  cross = FALSE,
  df = NULL,
  draws = NULL,
  eps = NULL,
  hypothesis = NULL,
  hypothesis_null = NULL,
  hypothesis_direction = NULL,
  jacobian = NULL,
  newdata = data.frame(),
  numderiv = list("fdforward"),
  type = NULL,
  variables = list(),
  vcov_model = NULL,
  wts = NULL
)
}
\arguments{
\item{model}{The fitted model object (required)}

\item{call}{The original function call (required)}

\item{df}{The degrees of freedom}

\item{newdata}{The new data frame for predictions}

\item{type}{The sanitized type from sanitize_type()}

\item{vcov_model}{The variance-covariance matrix}

\item{wts}{The weights specification}
}
\value{
An object of class marginaleffects_internal
}
\description{
Constructor for marginaleffects_internal class
}
\keyword{internal}
