% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.marginaleffects}
\alias{plot.marginaleffects}
\title{Plot average marginal effects}
\usage{
\method{plot}{marginaleffects}(mfx, conf.int = TRUE, conf.level = 0.95)
}
\description{
Plot average marginal effects
}
