% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slopes}
\alias{plot.slopes}
\title{Point-range plot of average marginal effects}
\usage{
\method{plot}{slopes}(x, conf_level = 0.95, ...)
}
\arguments{
\item{x}{An object produced by the \code{comparisons} function.}
}
\value{
A \code{ggplot2} object
}
\description{
Uses the \code{ggplot2} package to draw a point-range plot of the average marginal effects computed by \code{tidy}.
}
\details{
To compute standard errors around the average marginaleffects, we begin by applying the
mean function to each column of the Jacobian. Then, we use this matrix in the Delta
method to obtained standard errors.

In Bayesian models (e.g., \code{brms}), we compute Average Marginal
Effects by applying the mean function twice. First, we apply it to all
marginal effects for each posterior draw, thereby estimating one Average (or
Median) Marginal Effect per iteration of the MCMC chain. Second, we
calculate the mean and the \code{quantile} function to the results of Step 1 to
obtain the Average Marginal Effect and its associated interval.
}
\examples{
mod <- glm(am ~ hp + wt, data = mtcars)
mfx <- slopes(mod)
plot(mfx)

}
\seealso{
Other plot: 
\code{\link{plot_cap}()},
\code{\link{plot_cco}()},
\code{\link{plot_cme}()}
}
\concept{plot}
