% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cme.R
\name{plot_cme}
\alias{plot_cme}
\title{Plot Conditional Marginal Effects}
\usage{
plot_cme(
  model,
  effect = NULL,
  condition = NULL,
  type = "response",
  vcov = NULL,
  conf_level = 0.95,
  draw = TRUE,
  ...
)
}
\arguments{
\item{effect}{Name of the variable whose marginal effect we want to plot on the y-axis}

\item{condition}{character vector or named list of length smaller than 3. Character vectors must be the names of the predictor variables to display. The names of the list must The first element is displayed on the x-axis. The second element determines the colors. The third element creates facets. Unspecified variables are held at their means or modes. Lists can include these types of values (see Examples section below):
\itemize{
\item Numeric vector
\item Function which returns a numeric vector or a set of unique categorical values
\item Shortcut strings for common reference values: "minmax", "quartile", "threenum"
}}

\item{draw}{\code{TRUE} returns a \code{ggplot2} plot. \code{FALSE} returns a \code{data.frame} of the underlying data.}
}
\value{
A \code{ggplot2} object
}
\description{
This function plots marginal effects (y-axis) against values of predictor(s)
variable(s) (x-axis and colors). This is especially useful in models with
interactions, where the values of marginal effects depend on the values of
"condition" variables.
}
\examples{
library(marginaleffects)
mod <- lm(mpg ~ hp * drat * factor(am), data = mtcars)

plot_cme(mod, effect = "hp", condition = "drat")

plot_cme(mod, effect = "hp", condition = c("drat", "am"))

plot_cme(mod, effect = "hp", condition = list("am", "drat" = 3:5))

plot_cme(mod, effect = "am", condition = list("hp", "drat" = range))

plot_cme(mod, effect = "am", condition = list("hp", "drat" = "threenum"))

}
\seealso{
Other plot: 
\code{\link{plot.slopes}()},
\code{\link{plot_cap}()},
\code{\link{plot_cco}()}
}
\concept{plot}
