% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriordraws.R
\name{posteriordraws}
\alias{posteriordraws}
\title{Extract posterior draws from a \code{predictions} or \code{marginaleffects} object derived from Bayesian \code{brms} models}
\usage{
posteriordraws(x)
}
\arguments{
\item{x}{An object produced by the \code{predictions} or the \code{marginaleffects} functions}
}
\value{
A data.frame with \code{drawid} and \code{draw} columns.
}
\description{
Extract posterior draws from a \code{predictions} or \code{marginaleffects} object derived from Bayesian \code{brms} models
}
