% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_coefs.R
\name{reset_coefs}
\alias{reset_coefs}
\title{utility function to reset the coefficients in a model object, and return the modified model}
\usage{
reset_coefs(model, coefs)
}
\description{
utility function to reset the coefficients in a model object, and return the modified model
}
