% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R, R/get_dydx_betareg.R,
%   R/get_dydx_stats.R, R/get_dydx_nnet.R
\name{reset_coefs}
\alias{reset_coefs}
\alias{reset_coefs.default}
\alias{reset_coefs.betareg}
\alias{reset_coefs.glm}
\alias{reset_coefs.lm}
\alias{reset_coefs.multinom}
\title{Utility function to reset the coefficients in a model object}
\usage{
reset_coefs(model, coefs)

\method{reset_coefs}{default}(model, coefs)

\method{reset_coefs}{betareg}(model, coefs)

\method{reset_coefs}{glm}(model, coefs)

\method{reset_coefs}{lm}(model, coefs)

\method{reset_coefs}{multinom}(model, coefs)
}
\arguments{
\item{model}{object to modify}

\item{coefs}{vector of coefficients to insert in the model object}
}
\description{
Utility function to reset the coefficients in a model object
}
