% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_median_or_mode.R, R/get_dydx_stats.R,
%   R/get_dydx_nnet.R, R/get_dydx_MASS.R, R/get_dydx_betareg.R,
%   R/get_dydx_fixest.R, R/get_dydx_ivreg.R, R/get_dydx_lme4.R,
%   R/get_dydx_ordinal.R, R/get_dydx_survey.R
\name{reset_coefs}
\alias{reset_coefs}
\alias{reset_coefs.default}
\alias{reset_coefs.glm}
\alias{reset_coefs.lm}
\alias{reset_coefs.multinom}
\alias{reset_coefs.polr}
\alias{reset_coefs.betareg}
\alias{reset_coefs.fixest}
\alias{reset_coefs.ivreg}
\alias{reset_coefs.lmerMod}
\alias{reset_coefs.clm}
\alias{reset_coefs.svyglm}
\title{Utility function to reset the coefficients in a model object}
\usage{
reset_coefs(model, coefs)

\method{reset_coefs}{default}(model, coefs)

\method{reset_coefs}{glm}(model, coefs)

\method{reset_coefs}{lm}(model, coefs)

\method{reset_coefs}{multinom}(model, coefs)

\method{reset_coefs}{polr}(model, coefs)

\method{reset_coefs}{betareg}(model, coefs)

\method{reset_coefs}{fixest}(model, coefs)

\method{reset_coefs}{ivreg}(model, coefs)

\method{reset_coefs}{lmerMod}(model, coefs)

\method{reset_coefs}{clm}(model, coefs)

\method{reset_coefs}{svyglm}(model, coefs)
}
\arguments{
\item{model}{object to modify}

\item{coefs}{vector of coefficients to insert in the model object}
}
\description{
Utility function to reset the coefficients in a model object
}
