% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_model.R, R/methods_nnet.R,
%   R/methods_brms.R, R/methods_mclogit.R, R/methods_mlogit.R, R/methods_plm.R,
%   R/methods_quantreg.R
\name{sanity_model_specific}
\alias{sanity_model_specific}
\alias{sanity_model_specific.default}
\alias{sanity_model_specific.multinom}
\alias{sanity_model_specific.brmsfit}
\alias{sanity_model_specific.mblogit}
\alias{sanity_model_specific.mlogit}
\alias{sanity_model_specific.plm}
\alias{sanity_model_specific.rqs}
\title{Method to raise model-specific warnings and errors}
\usage{
sanity_model_specific(model, calling_function = "marginaleffects", ...)

\method{sanity_model_specific}{default}(model, calling_function = "marginaleffects", ...)

\method{sanity_model_specific}{multinom}(model, calling_function = "marginaleffects", ...)

\method{sanity_model_specific}{brmsfit}(model, ...)

\method{sanity_model_specific}{mblogit}(model, calling_function = "marginaleffects", ...)

\method{sanity_model_specific}{mlogit}(model, newdata, ...)

\method{sanity_model_specific}{plm}(model, ...)

\method{sanity_model_specific}{rqs}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are passed to the \code{predict()} method used to
compute adjusted predictions. These arguments are particularly useful for
mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?marginaleffects} documentation for a non-exhaustive list of available
arguments.}

\item{newdata}{\code{NULL}, a data frame or a string which determines the predictor values for which to compute marginaleffects.
\itemize{
\item \code{NULL} (default): Unit-level marginal effects for each observed value in the original dataset.
\item A data frame: Unit-level marginal effects for each row of the \code{newdata} data frame.
\item "mean": Marginal Effects at the Mean. Marginal effects when each predictor is held at its mean or mode.
\item "median": Marginal Effects at the Median. Marginal effects when each predictor is held at its mean or mode.
\item "marginalmeans": Marginal Effects at Marginal Means. See Details section below.
\item The \code{\link[=datagrid]{datagrid()}} function can be used to specify a custom grid of regressors. For example:
\itemize{
\item \code{newdata = datagrid(cyl = c(4, 6))}: \code{cyl} variable equal to 4 and 6 and other regressors fixed at their means or modes.
\item See the Examples section and the \code{\link[=datagrid]{datagrid()}} documentation.
}
}}
}
\value{
A warning, an error, or nothing
}
\description{
Method to raise model-specific warnings and errors
}
\keyword{internal}
