% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_model.R, R/methods_nnet.R,
%   R/methods_plm.R, R/methods_quantreg.R, R/methods_robust.R
\name{sanity_model_specific}
\alias{sanity_model_specific}
\alias{sanity_model_specific.default}
\alias{sanity_model_specific.multinom}
\alias{sanity_model_specific.plm}
\alias{sanity_model_specific.rqs}
\alias{sanity_model_specific.lmRob}
\title{Method to raise model-specific warnings and errors}
\usage{
sanity_model_specific(model, ...)

\method{sanity_model_specific}{default}(model, ...)

\method{sanity_model_specific}{multinom}(...)

\method{sanity_model_specific}{plm}(model, ...)

\method{sanity_model_specific}{rqs}(model, ...)

\method{sanity_model_specific}{lmRob}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
A warning, an error, or nothing
}
\description{
Method to raise model-specific warnings and errors
}
\keyword{internal}
