% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_coef.R, R/methods_betareg.R,
%   R/methods_nnet.R, R/methods_pscl.R, R/methods_stats.R
\name{set_coef}
\alias{set_coef}
\alias{set_coef.default}
\alias{set_coef.betareg}
\alias{set_coef.multinom}
\alias{set_coef.hurdle}
\alias{set_coef.glm}
\alias{set_coef.lm}
\title{Internal function to set coefficients

Set the coefficients in a model to different values and return the modified object (internal function)}
\usage{
set_coef(model, coefs)

\method{set_coef}{default}(model, coefs)

\method{set_coef}{betareg}(model, coefs)

\method{set_coef}{multinom}(model, coefs)

\method{set_coef}{hurdle}(model, coefs)

\method{set_coef}{glm}(model, coefs)

\method{set_coef}{lm}(model, coefs)
}
\arguments{
\item{model}{object to modify}

\item{coefs}{vector of coefficients to insert in the model object}
}
\description{
Internal function to set coefficients

Set the coefficients in a model to different values and return the modified object (internal function)
}
\keyword{internal}
