% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.comparisons}
\alias{summary.comparisons}
\title{Summarize a \code{comparisons} object}
\usage{
\method{summary}{comparisons}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{An object produced by the \code{comparisons} function}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int=TRUE}. Must be strictly greater than 0 and less than 1. Defaults
to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments are passed to the \code{predict()} method used to
compute adjusted predictions. These arguments are particularly useful for
mixed-effects or bayesian models (see the online vignettes on the
\code{marginaleffects} website). Available arguments can vary from model to
model, depending on the range of supported arguments by each modeling
package. See the "Model-Specific Arguments" section of the
\code{?marginaleffects} document for a non-exhaustive list of available
arguments.}
}
\value{
Data frame of summary statistics for an object produced by the
\code{comparisons} function
}
\description{
Summarize a \code{comparisons} object
}
