% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.marginalmeans}
\alias{summary.marginalmeans}
\title{Summarize a \code{marginalmeans} object}
\usage{
\method{summary}{marginalmeans}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{An object produced by the \code{marginalmeans} function}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int=TRUE}. Must be strictly greater than 0 and less than 1. Defaults
to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
Data frame of summary statistics for an object produced by the
\code{marginalmeans} function
}
\description{
Summarize a \code{marginalmeans} object
}
