% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.predictions}
\alias{summary.predictions}
\title{Summarize a \code{predictions} object}
\usage{
\method{summary}{predictions}(object, ...)
}
\arguments{
\item{object}{An object produced by the \code{predictions} function}

\item{...}{The "Model-Specific Arguments" section below gives a list of arguments which can modify the behavior of this function for certain models (e.g., mixed-effects or bayesian).}
}
\value{
Data frame of summary statistics for an object produced by the
\code{predictions} function
}
\description{
Summarize a \code{predictions} object
}
