% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.predictions}
\alias{summary.predictions}
\title{Summarize a \code{predictions} object}
\usage{
\method{summary}{predictions}(object, ...)
}
\arguments{
\item{object}{An object produced by the \code{predictions} function}

\item{...}{Additional arguments are pushed forward to \code{predict()}.}
}
\value{
Data frame of summary statistics for an object produced by the
\code{predictions} function
}
\description{
Summarize a \code{predictions} object
}
