% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_models.R
\docType{data}
\name{supported_models}
\alias{supported_models}
\title{List of model types supported by the \code{marginaleffects} package}
\format{
data.frame
\describe{
\item{Package}{R package which produces the model object.}
\item{Function}{R function which produces the model object.}
\item{dydx}{Can \code{marginaleffects} retrieve marginal effects?}
\item{se}{Can \code{marginaleffects} retrieve standard errors?}
\item{stata_dydx}{Has the validity of numerical results for marginal effects been checked against Stata's margins command?}
\item{stata_se}{Has the validity of numerical results for standard errors been checked against Stata's margins command?}
\item{stata_dydx}{Has the validity of numerical results for marginal effects been checked against the margins package for R??}
\item{stata_se}{Has the validity of numerical results for standard errors been checked against the margins package for R??}
}
}
\usage{
supported_models
}
\description{
A list of model types supported by the \code{marginaleffects} package
}
\keyword{datasets}
