% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{typical}
\alias{typical}
\title{Superseded by datagrid(...)}
\usage{
typical(
  ...,
  model = NULL,
  newdata = NULL,
  FUN_character = Mode,
  FUN_factor = Mode,
  FUN_logical = Mode,
  FUN_numeric = function(x) mean(x, na.rm = TRUE),
  FUN_other = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{...}{named arguments with vectors of values for user-specified
variables. The output will include all combinations of these variables (see
Examples below.)}

\item{model}{Model object}

\item{newdata}{data.frame (one and only one of the \code{model} and \code{newdata} arguments}

\item{FUN_character}{the function to be applied to character variables.}

\item{FUN_factor}{the function to be applied to factor variables.}

\item{FUN_logical}{the function to be applied to factor variables.}

\item{FUN_numeric}{the function to be applied to numeric variables.}

\item{FUN_other}{the function to be applied to other variable types.}
}
\description{
Superseded by datagrid(...)
}
\keyword{internal}
